### 2024-05-30 #6813 @editorialbot
[PRE REVIEW]: SegregatedVMSSolver.jl: Linearized and Segregated Stabilized Solver for Large Eddy Simulation in Julia
**Submitting author:** @carlodev (<a href="http://orcid.org/0000-0002-2873-6293">Carlo Brunelli</a>)
**Repository:** https://github.com/carlodev/SegregatedVMSSolver.jl
**Branch with paper.md** (empty if default branch): JossPaper
**Version:** v2.0.0
