### 2024-06-25 3321c92
add SyntheticEddyMethod interface support
