+++ src/Exports.jl
@@ -20,6 +20,7 @@ end
 
 @publish ParametersDef TimeParameters
 @publish ParametersDef PhysicalParameters
+@publish ParametersDef TurbulenceParameters
 @publish ParametersDef SolverParameters
 @publish ParametersDef MeshParameters
 @publish ParametersDef CartesianMeshParams
