+++ Project.toml
@@ -22,6 +22,7 @@ Revise = "295af30f-e4ad-537b-8983-00126c2a3abe"
 ScatteredInterpolation = "3f865c0f-6dca-5f4d-999b-29fe1e7e3c92"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
+SyntheticEddyMethod = "57cf2c78-4fcb-4d27-9f1f-369b8b08b789"
 Trapz = "592b5752-818d-11e9-1e9a-2b8ca4a44cd1"
 UnPack = "3a884ed6-31ef-47d7-9d2a-63182c4928ed"
 

[…]
