### 2024-06-25 f366b4d
bug fix - add intialization SimulationParameters
