### 2024-06-25 cef8b31
do not compute fluctuations while ramping
