### 2024-06-25 348e7f2
reduce number of eddies
