+++ src/Commons/ParametersDef/Cases.jl
@@ -1,4 +1,5 @@
 abstract type SimulationCase end
+abstract type VelocityBoundaryCase<:SimulationCase end
 
 struct SimulationParameters
     timep::TimeParameters
@@ -6,39 +7,50 @@ struct SimulationParameters
     turbulencep::TurbulenceParameters
     solverp::SolverParameters
     exportp::ExportParameters
-    restartp::RestartParameters
+    intialp::InitialParameters
 end
 
 function SimulationParameters(timep::TimeParameters,physicalp::PhysicalParameters,
                             solverp::SolverParameters,exportp::ExportParameters)
     turbulencep=TurbulenceParameters()
-    restartp=RestartParameters()
-    SimulationParameters(timep,physicalp,turbulencep,solverp,exportp,restartp)
+    intialp=InitialParameters()
+    SimulationParameters(timep,physicalp,turbulencep,solverp,exportp,intialp)
 end
 
 function SimulationParameters(timep::TimeParameters,physicalp::PhysicalParameters, turbulencep::TurbulenceParameters,
     solverp::SolverParameters,exportp::ExportParameters)
-    restartp=RestartParameters()
-    SimulationParameters(timep,physicalp,turbulencep,solverp,exportp,restartp)
+    intialp=InitialParameters()
+    SimulationParameters(timep,physicalp,turbulencep,solverp,exportp,intialp)
 end
 
 function SimulationParameters(timep::TimeParameters,physicalp::PhysicalParameters,
-    solverp::SolverParameters,exportp::ExportParameters, restartp::RestartParameters)
+    solverp::SolverParameters,exportp::ExportParameters, intialp::InitialParameters)
     turbulencep=TurbulenceParameters()
-    SimulationParameters(timep,physicalp,turbulencep,solverp,exportp,restartp)
+    SimulationParameters(timep,physicalp,turbulencep,solverp,exportp,intialp)
 end
 
+"""
+    create_new_case(case::Symbol)
 
-for case in (:Airfoil,:WindTunnel,:Cylinder,:LidDriven)
+It creates a new case, named `case`. It is useful to add new simulation cases in a high level way.
+"""
+function create_new_case(case::Symbol)
     @eval begin
-        struct $case <: SimulationCase
+        struct $case <: VelocityBoundaryCase
             meshp::MeshParameters
             simulationp::SimulationParameters
             sprob::StabilizedProblem
         end
+        export $case
     end
 end
-  
+
+for case in (:Airfoil,:WindTunnel,:Cylinder,:LidDriven)
+    create_new_case(case::Symbol)
+end
+
+
+
 struct TaylorGreen <: SimulationCase
     analyticalsol
     meshp::MeshParameters
@@ -59,7 +71,11 @@ function TaylorGreen(meshp::MeshParameters, simulationp::SimulationParameters, s
     TaylorGreen(analyticalsol, meshp, simulationp, sprob)
 end
 
-VelocityBoundaryCase = Union{Airfoil,WindTunnel,Cylinder, LidDriven}
+
+
+
+# VelocityBoundaryCase = Union{Airfoil,WindTunnel,Box,Cylinder, LidDriven}
+
 
 
 MyStructurePrint = Union{SimulationCase,StabilizedProblem,SimulationParameters,
@@ -135,17 +151,21 @@ end
 For the point x, at the time t it computes the velocity fluctuations in all the direction. Each time the time t is increased the Eddy are convected.
 The time informations are coded in the VirtualBox.
 """
[…]
