### 2024-07-01 c2baa99
add compute_CL_CD_time support
