### 2024-07-02 bf6041b
add condition SEM
