### 2024-07-27 1c1f351
CompatHelper: bump compat for PartitionedArrays to 0.5, (keep existing compat)
