+++ Project.toml
@@ -37,7 +37,7 @@ GridapPETSc = "0.5.0"
 MPI = "0.20.16"
 NearestNeighbors = "0.4.16"
 Parameters = "0.12.3"
-PartitionedArrays = "0.3.4"
+PartitionedArrays = "0.3.4, 0.5"
 Revise = "3.5.9"
 ScatteredInterpolation = "0.3.6"
 Trapz = "2.0.3"
