### 2024-09-26 fce6dd5
natural and periodic TGV case
