+++ test/TestsCases/TestsCases.jl
@@ -11,7 +11,8 @@ include("TurbulentAirfoilTest.jl")
 include("AirfoilTest.jl")
 include("CylinderTest.jl")
 include("LidDrivenTest.jl")
-include("TaylorGreenTest.jl")
+include("TGV_Natural.jl")
+include("TGV_Periodic.jl")
 
 function tests_cases(backend)
   backend() do distribute
@@ -32,9 +33,14 @@ function tests_cases(backend)
     airfoil_test(backend)
     cylinder_test(backend)
     liddriven_test(backend)
-    taylorgreen_test(backend)
+    TGV_Periodic_test(backend)
+    TGV_Natural_test(backend)
   end
 
 end
+
+
+
+tests_cases(with_debug)
 
 end #end_module
