+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"mpi/#MPI-Run","page":"MPI run","title":"MPI Run","text":"","category":"section"},{"location":"mpi/","page":"MPI run","title":"MPI run","text":"The code can be run in Message Passing Interface (MPI).  The code is made in such a way that it can run:","category":"page"},{"location":"mpi/","page":"MPI run","title":"MPI run","text":"in the REPL, selecting SegregatedVMSSolver.solve(simcase,with_debug)\nin MPI, selecting SegregatedVMSSolver.solve(simcase,with_mpi)","category":"page"},{"location":"mpi/","page":"MPI run","title":"MPI run","text":"When running in MPI the code cannot be easily executed in the REPL. Instead, one has to run them from a terminal using the mpiexecjl script as provided by MPI.jl. ","category":"page"},{"location":"mpi/","page":"MPI run","title":"MPI run","text":"For example with the command: ","category":"page"},{"location":"mpi/","page":"MPI run","title":"MPI run","text":"mpiexecjl -n 4 julia --project=. run_simulation.jl","category":"page"},{"location":"mpi/","page":"MPI run","title":"MPI run","text":"It is possible to launch the tests changing the backend.","category":"page"},{"location":"mpi/#Run-in-a-Cluster","page":"MPI run","title":"Run in a Cluster","text":"","category":"section"},{"location":"mpi/","page":"MPI run","title":"MPI run","text":"It is possible to run a simulation in a cluster, creating a suitable bash file, example the following run_sim.sh file can be used. The user has to specify where julia is installed and to activate a suitable project where there is SegregatedVMSSolver in the dependancies. The run_simulation.jl is the julia file of the simulation itself. Be sure that the number of processors specified in simulation is corresponding to the number of processors used for running the simulation in the cluster.","category":"page"},{"location":"mpi/","page":"MPI run","title":"MPI run","text":"#!/bin/sh\nexport PATH=$HOME/julia-1.8.5/bin/:$PATH\n\nsrun julia --project=../../../ -O3 --check-bounds=no -L run_simulation.jl","category":"page"},{"location":"mpi/","page":"MPI run","title":"MPI run","text":"This is a command example to launch the simulation on a single node (-N1) a cluster using slurm, using 80 cores -n80, specifing the partition of 80 cores -p80CORE and a total time of 400 hours -t 400:0:0.","category":"page"},{"location":"mpi/","page":"MPI run","title":"MPI run","text":"sbatch -t 400:0:0 -N 1 -n80 -p80CORE run_sim.sh","category":"page"},{"location":"references/#References","page":"References","title":"References","text":"","category":"section"},{"location":"references/","page":"References","title":"References","text":"","category":"page"},{"location":"user_params/#User-Parameters","page":"User Parameters","title":"User Parameters","text":"","category":"section"},{"location":"user_params/","page":"User Parameters","title":"User Parameters","text":"The simulation parameters are written step by step by the user. The Physical Parameters are defined in src/Commons/ParametersDef/Params.jl, where the default values are also provided.","category":"page"},{"location":"user_params/#TimeParameters","page":"User Parameters","title":"TimeParameters","text":"","category":"section"},{"location":"user_params/","page":"User Parameters","title":"User Parameters","text":"using SegregatedVMSSolver\nusing PartitionedArrays,MPI\nusing SegregatedVMSSolver.ParametersDef\nt0 = 0.0\ndt = 0.1\ntF = 1.0\n\ntimep = TimeParameters(t0,dt,tF)","category":"page"},{"locati
[…]
