### 2024-09-27 3cf9ecc
CompatHelper: add new compat entry for SyntheticEddyMethod at version 0.4, (keep existing compat)
