+++ Project.toml
@@ -43,6 +43,7 @@ Parameters = "0.12.3"
 PartitionedArrays = "0.3.4"
 Revise = "3.5.9"
 ScatteredInterpolation = "0.3.6"
+SyntheticEddyMethod = "0.4"
 Trapz = "2.0.3"
 julia = "1.8.2"
 
