### 2024-09-27 5358507
CompatHelper: add new compat entry for FileIO at version 1, (keep existing compat)
