+++ test/TestsCases/TGV_Natural.jl
@@ -3,16 +3,20 @@ using SegregatedVMSSolver
 using SegregatedVMSSolver.ParametersDef
 using SegregatedVMSSolver.SolverOptions
 using MPI
+using Test
 
-function TGV_Natural_test(backend)
+# function TGV_Natural_test(backend)
+backend = with_debug
     t0 =0.0
     dt = 0.1
     tF = 3.0
     vortex_diameter = 1.0
-
+    
+    N = 16
+    
     Re = 500
-    D = 3
-    rank_partition = (2,2,2)
+    D = 2
+    rank_partition = (2,2)
 
 
     sprob = StabilizedProblem(VMS(1))
@@ -20,10 +24,10 @@ function TGV_Natural_test(backend)
 
     physicalp = PhysicalParameters(Re=Re,c=vortex_diameter)
     solverp = SolverParameters()
-    exportp = ExportParameters(printinitial=true,printmodel=true)
+    exportp = ExportParameters(printinitial=false,printmodel=false)
 
 
-    meshp= MeshParameters(rank_partition,D;N=32,L=vortex_diameter/2)
+    meshp= MeshParameters(rank_partition,D;N=N,L=vortex_diameter/2)
     simparams = SimulationParameters(timep,physicalp,solverp,exportp)
 
     bc_tgv = Natural(meshp,physicalp ) 
@@ -35,6 +39,6 @@ function TGV_Natural_test(backend)
 
     @test SegregatedVMSSolver.solve(mcase,backend)
 
-end
+# end
 
 
