+++ test/TestsCases/TGV_Natural.jl
@@ -7,7 +7,7 @@ using MPI
 function TGV_Natural_test(backend)
     t0 =0.0
     dt = 0.1
-    tF = 3.0
+    tF = dt * 21
     vortex_diameter = 1.0
 
     Re = 500
@@ -23,7 +23,7 @@ function TGV_Natural_test(backend)
     exportp = ExportParameters(printinitial=true,printmodel=true)
 
 
-    meshp= MeshParameters(rank_partition,D;N=32,L=vortex_diameter/2)
+    meshp= MeshParameters(rank_partition,D;N=16,L=vortex_diameter/2)
     simparams = SimulationParameters(timep,physicalp,solverp,exportp)
 
     bc_tgv = Natural(meshp,physicalp ) 
