### 2024-09-30 b45d98d
output Ek, Enstrophy for Natural
