### 2024-10-01 e55d849
CompatHelper: add new compat entry for DelimitedFiles at version 1, (keep existing compat)
