+++ Project.toml
@@ -32,6 +32,7 @@ UnPack = "3a884ed6-31ef-47d7-9d2a-63182c4928ed"
 [compat]
 CSV = "0.10.11"
 DataFrames = "1.6.1"
+DelimitedFiles = "1"
 FFTW = "1.8.0"
 FillArrays = "1.7.0"
 Gridap = "0.17.20"
