+++ previews/PR23/docs.bib
@@ -1,0 +1,70 @@
+@article{Bazilevs2007,
+   author = {Y. Bazilevs and V. M. Calo and J. A. Cottrell and T. J.R. Hughes and A. Reali and G. Scovazzi},
+   doi = {10.1016/j.cma.2007.07.016},
+   issn = {00457825},
+   issue = {1-4},
+   journal = {Computer Methods in Applied Mechanics and Engineering},
+   keywords = {Homogeneous isotropic turbulence,Incompressible flows,Isogeometric analysis,Large eddy simulation,NURBS,Optimality,Perturbation series,Projection,Small-scale Green's function,Turbulence modeling,Turbulent channel flows,Variational multiscale methods},
+   month = {12},
+   pages = {173-201},
+   title = {Variational multiscale residual-based turbulence modeling for large eddy simulation of incompressible flows},
+   volume = {197},
+   year = {2007},
+}
+
+@phdthesis{Banyai2016,
+   author = {Tamas Banyai},
+   isbn = {9782875161093},
+   school = {Université Libre de Bruxelles},
+   title = {Development of Stabilized Finite Element Method for Numerical Simulation of Turbulent Incompressible Single and Eulerian-Eulerian Two-Phase Flows},
+   year = {2016},
+}
+
+@phdthesis{Janssens2014,
+   author = {Bart Janssens},
+   isbn = {978-2-87516-081-2},
+   school = {Université Libre de Bruxelles},
+   title = {Numerical modeling and experimental investigation of ﬁne particle coagulation
+   and dispersion in dilute ﬂows},
+   year = {2014},
+}
+
+
+@misc{Verdugo2021,
+   author = {Francesc Verdugo and Santiago Badia},
+   doi = {10.1016/j.cpc.2022.108341},
+   month = {9},
+   title = {The software design of Gridap: a Finite Element package based on the Julia JIT compiler},
+   url = {http://arxiv.org/abs/2109.12818 http://dx.doi.org/10.1016/j.cpc.2022.108341},
+   year = {2021},
+}
+
+@misc{Hughes2000,
+   author = {Thomas J R Hughes and Luca Mazzei and Kenneth E Jansen},
+   journal = {Comput Visual Sci},
+   pages = {47-59},
+   title = {Computing and Visualization in Science Large Eddy Simulation and the variational multiscale method},
+   volume = {3},
+   year = {2000},
+}
+@article{Trofimova2009,
+   author = {Alisa V. Trofimova and Andrés E. Tejada-Martínez and Kenneth E. Jansen and Richard T. Lahey},
+   doi = {10.1016/j.compfluid.2008.10.003},
+   issn = {00457930},
+   issue = {4},
+   journal = {Computers and Fluids},
+   month = {4},
+   pages = {924-938},
+   title = {Direct numerical simulation of turbulent channel flows using a stabilized finite element method},
+   volume = {38},
+   year = {2009},
+}
+
+@misc{Bakker2018,
+   author = {J C Bakker},
+   title = {Wall-Distance Calcu-lation for Turbulence Modelling},
+   url = {http://repository.tudelft.nl/.},
+    year={2018},
+    school = {Delft University of Technology},
+    type     = {Bacherlor's thesis},
+}
