### 2024-11-15 #6813 @Kevin-Mattheus-Moerman
@fredrikekre @termi-official @santiagobadia @Balaje @weymouth @maeckha @ali-ramadhan would you be interested in reviewing this submission for [JOSS](https://joss.theoj.org/) entitled: "SegregatedVMSSolver.jl: Linearized and Segregated Stabilized Solver for Large Eddy Simulation in Julia"?. The work uses Julia and GridAp for solving fluid dynamics problems.

[JOSS reviews](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) take place here on GitHub, and focus on [the software](https://github.com/carlodev/SegregatedVMSSolver.jl) as well [a short paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.06813/joss.06813/10.21105.joss.06813.pdf).

If you are interested please let me know here. Thanks!
