### 2024-12-20 744a93d
add TGV Periodic 3D support
