+++ test/TestsCases/TGV_Natural.jl
@@ -5,32 +5,39 @@ using SegregatedVMSSolver.SolverOptions
 using MPI
 using Test
 
+
 function TGV_Natural_test(backend)
+    @test TGV_Natural_test(backend,2)
+    @test TGV_Natural_test(backend,3)
+end
+
+
+function TGV_Natural_test(backend, D::Int64)
+
     t0 =0.0
-    dt = 0.1
-    tF = dt * 21
+    dt = 1e-2
+    tF = dt * 3
     vortex_diameter = 1.0
     
-    N = 16
+    N = 32
     
     Re = 500
-    D = 2
-    rank_partition = (2,2)
+    rank_partition =  ntuple(i -> 2, D)
 
 
     sprob = StabilizedProblem(VMS(1))
     timep = TimeParameters(t0=t0,dt=dt,tF=tF)
 
     physicalp = PhysicalParameters(Re=Re,c=vortex_diameter)
-    solverp = SolverParameters()
-    exportp = ExportParameters(printinitial=false,printmodel=false)
+    solverp = SolverParameters(matrix_freq_update = 1, M=2)
+    exportp = ExportParameters(printinitial=true,printmodel=true)
 
 
-    meshp= MeshParameters(rank_partition,D;N=16,L=vortex_diameter/2)
+    meshp= MeshParameters(rank_partition,D;N=N,L=vortex_diameter/2)
 
     simparams = SimulationParameters(timep,physicalp,solverp,exportp)
 
-    bc_tgv = Natural(meshp,physicalp ) 
+    bc_tgv = Natural(meshp,physicalp) 
 
 
     mcase = TaylorGreen(bc_tgv, meshp,simparams,sprob)
