+++ src/Commons/CreateProblem/ModelCreation.jl
@@ -73,23 +73,29 @@ end
 
 function create_model(parts, simcase::TaylorGreen{Periodic})
     mesh = simcase.meshp
-    rank_partition = mesh.rank_partition
+    @unpack rank_partition,D, meshinfo = mesh
 
-    domain,partition = create_model(mesh.meshinfo, mesh.D)
+    domain,partition = create_model(meshinfo, mesh.D)
 
-    model =CartesianDiscreteModel(parts,rank_partition, domain, partition;isperiodic=(true, true) )
-    model = add_centre_tag!(model, Point(0.0, 0.0)) #(0.0, 0.0) is the centre coordinate
+    model =CartesianDiscreteModel(parts,rank_partition, domain, partition;isperiodic=simcase.bc_type.bc )
+    if D == 2
+        model = add_centre_tag!(model, Point(0.0, 0.0)) #(0.0, 0.0) is the centre coordinate
+    end
+
     
     print_model(model,simcase)
     return model 
 end
 
 
+
+
+
 function create_model(parts, simcase::TaylorGreen{Natural})
     mesh = simcase.meshp
     rank_partition = mesh.rank_partition
     D = mesh.D
-    
+    L = mesh.meshinfo.L
     domain,partition = create_model(mesh.meshinfo, D)
 
     model =CartesianDiscreteModel(parts,rank_partition, domain, partition)
@@ -99,6 +105,7 @@ function create_model(parts, simcase::TaylorGreen{Natural})
 
     if D == 2
         add_tag_from_tags!(labels, "external", collect(1:8))
+
     elseif D == 3
         add_tag_from_tags!(labels, "external", collect(1:26))
     end
