+++ src/Commons/CreateProblem/SpaceConditions.jl
@@ -1,4 +1,33 @@
-function creation_fe_spaces(simcase::TaylorGreen, model, boundary_conditions)
+function creation_fe_spaces(simcase::TaylorGreen{Periodic}, model, boundary_conditions)
+
+    u_diri_tags,u_diri_values,p_diri_tags,p_diri_values = boundary_conditions
+
+    @sunpack D,order = simcase
+    reffeᵤ = ReferenceFE(lagrangian, VectorValue{D, Float64}, order)
+    reffeₚ = ReferenceFE(lagrangian, Float64, order)
+
+
+    V = TestFESpace(model, reffeᵤ, conformity=:H1)
+    U = TransientTrialFESpace(V)
+    if D == 2
+        Q = TestFESpace(model, reffeₚ, conformity=:H1, dirichlet_tags=p_diri_tags)
+        P = TransientTrialFESpace(Q, p_diri_values)
+    elseif D ==3
+        Q = TestFESpace(model, reffeₚ, conformity=:H1)
+        P = TransientTrialFESpace(Q)        
+    end
+
+    
+
+    Y = MultiFieldFESpace([V, Q])
+    X = TransientMultiFieldFESpace([U, P])
+
+    return V, U, P, Q, Y, X
+
+end
+
+
+function creation_fe_spaces(simcase::TaylorGreen{Natural}, model, boundary_conditions)
 
     u_diri_tags,u_diri_values,p_diri_tags,p_diri_values = boundary_conditions
 

[…]
