+++ Project.toml
@@ -40,5 +40,5 @@ Random = "1.11.0"
 Revise = "3.6.4"
 SparseArrays = "1.11.0"
 Statistics = "1.11.1"
-SyntheticEddyMethod = "0.4.2"
+SyntheticEddyMethod = "0.4.3"
 UnPack = "1.0.2"
