### 2025-01-08 carlodev/SegregatedVMSSolver.jl#24 @marinlauber
Test breaking
Hi! I am adding this issue while working on the review (https://github.com/openjournals/joss-reviews/issues/7564).

The installation of the software works as expected, but 2 of the test fail on my machine with `julia 1.11.1`.

```
    Test Summary:                       | Pass  Error  Total      Time
    Sequential                          |   32      2     34  54m03.4s
    Sequential                          |   32      2     34  53m50.9s
        Commons Sequential              |   14      2     16   1m53.0s
        Create Problem with_debug       |    9      2     11     52.3s
            New Tag                     |    1             1      0.1s
        Test Solve Problem with_debug   |    5             5     52.6s
        Cases Tests Sequential          |   18            18  51m57.9s
    ERROR: LoadError: Some tests did not pass: 32 passed, 0 failed, 2 errored, 0 broken
```
 Also, These are very slow to run, it might be best to split them to unit tests and actual verification/validation tests, or simplify the verification/validation tests.
