### 2025-01-08 carlodev/SegregatedVMSSolver.jl#25 @marinlauber
Missing Community Guidelines
I am adding this issue while working on the review (https://github.com/openjournals/joss-reviews/issues/7564).

The repository is missing **Community Guidelines**, which are requested as part of the review process.
