### 2025-01-08 carlodev/SegregatedVMSSolver.jl#26 @marinlauber
Add dedicated examples folder
This issue is related to https://github.com/openjournals/joss-reviews/issues/7564

There are a few _examples_ in the tests folder, but tits unclear how to run them. I think potential users will be grateful to have a dedicated `examples` folder, with either scripts that can be run interactively in the `REPL` or notebooks that detail how to set up a problem and plot the solution.

These can be the setups that generate the figures shown in the paper, which additionally would help make the results reproducible.
