### 2025-01-08 #7564 @marinlauber
Thank you @carlodev for your software.

Could you please address the following in the paper and the documentation?

- [ ] In the **Statement of Needs**, there is repeated information (lines 41 to 46), and it is not entirely clear which case is 3D and which one is 2D. I suggest to rephrase the text to improve readability.
- [ ] If the focus of the package is in solving the flow over 2D/3D airfoils, it should be made clear at the start of the statement of needs, here it is briefly mentioned in the 3rd paragraph.
 - [ ] **Reproducibility**, I am not sure how to reproduce the results presented in Figure 1.
 - [ ] **Functionality** and **Performance**, because there is no clear way of running some examples, I am not sure if all the functionalities and performance of the code match. From the tests, it seems that the code runs fine on multiprocessors.
 - [ ] **State of the field**, there is no clear indication on the state of the field or common package used for this.
 - [ ] **Citations**, the citations on line 59-60 should probably be changed to `\citet{}` to avoid the repeated name in the text.

I also opened a few issues  ([#24](https://github.com/carlodev/SegregatedVMSSolver.jl/issues/24), [#25](https://github.com/carlodev/SegregatedVMSSolver.jl/issues/25), [#26](https://github.com/carlodev/SegregatedVMSSolver.jl/issues/26)) on the source repository that complete my review.

Thank you in advance for considering my suggestions

Marin
