### 2025-01-10 carlodev/SegregatedVMSSolver.jl#27 @nathanquinlan
`ParametersDef` not defined in `SegregatedVMSSolver` when running example cases
Hi @carlodev,

Raising this issue in the course of a [JOSS review](https://github.com/openjournals/joss-reviews/issues/7564). I am unable to run the Taylor-Green flow following [instructions](https://carlodev.github.io/SegregatedVMSSolver.jl/dev/Cases/liddriven/). Similarly for other cases. I get an error as follows:
```julia
julia> using PartitionedArrays

julia> using SegregatedVMSSolver

julia> using SegregatedVMSSolver.ParametersDef
ERROR: UndefVarError: `ParametersDef` not defined in `SegregatedVMSSolver`
Suggestion: check for spelling errors or missing imports.
```
