### 2025-01-13 cddb235
rm turbulent_arifoil_test
