+++ test/TestsCases/TestsCases.jl
@@ -28,7 +28,6 @@ function tests_cases(backend)
   end
 
   @testset "Cases Tests $(typeof(backend))" begin
-    turbulent_airfoil_test(backend, true)
     turbulent_airfoil_test(backend, false)
     airfoil_test(backend)
     cylinder_test(backend)
