### 2025-01-13 8a428a6
fix turbulent parameters test
