+++ docs/make.jl
@@ -6,7 +6,7 @@ using Documenter, DocumenterCitations, SegregatedVMSSolver
 bib = CitationBibliography(joinpath(@__DIR__, "src", "docs.bib"); style=:numeric)
 
 
-makedocs(bib;
+makedocs(;plugins=[bib],
     sitename = "SegregatedVMSSolver.jl",
     modules = [SegregatedVMSSolver],
     pages = [
@@ -20,7 +20,6 @@ makedocs(bib;
         "Lid Driven Cavity Flow" => "Cases/liddriven.md",
         "Cylinder" => "Cases/cylinder.md",
         ],
-        "Restart" => "restart.md",
 
         "Post Processing" => "post_proc.md",
         "Boundary Layer Initialization" => "blinit.md",
