+++ Project.toml
@@ -33,12 +33,10 @@ UnPack = "3a884ed6-31ef-47d7-9d2a-63182c4928ed"
 [compat]
 DataInterpolations = "6.6.0"
 DelimitedFiles = "1.9.1"
-FileIO = "1.16.6"
-GridapPETSc = "0.5.3"
 LinearAlgebra = "1.11.0"
-Random = "1.11.0"
-Revise = "3.6.4"
 SparseArrays = "1.11.0"
-Statistics = "1.11.1"
-SyntheticEddyMethod = "0.4.3"
 UnPack = "1.0.2"
+GridapPETSc = "0.5.3"
+Gridap = "0.17.20"
+GridapDistributed ="0.3.2"
+GridapGmsh="0.7.1"
