+++ .github/workflows/CI.yml
@@ -13,7 +13,7 @@ jobs:
       fail-fast: false
       matrix:
         version:
-          - '1.8.2'
+          - '1.11.1'
         os:
           - ubuntu-latest
         arch:
