+++ Project.toml
@@ -27,16 +27,18 @@ ScatteredInterpolation = "3f865c0f-6dca-5f4d-999b-29fe1e7e3c92"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 SyntheticEddyMethod = "57cf2c78-4fcb-4d27-9f1f-369b8b08b789"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 Trapz = "592b5752-818d-11e9-1e9a-2b8ca4a44cd1"
 UnPack = "3a884ed6-31ef-47d7-9d2a-63182c4928ed"
 
 [compat]
 DataInterpolations = "6.6.0"
 DelimitedFiles = "1.9.1"
+Gridap = "0.17.20"
+GridapDistributed = "0.3.2"
+GridapGmsh = "0.7.1"
+GridapPETSc = "0.5.3"
 LinearAlgebra = "1.11.0"
 SparseArrays = "1.11.0"
+Test = "1.11.0"
 UnPack = "1.0.2"
-GridapPETSc = "0.5.3"
-Gridap = "0.17.20"
-GridapDistributed ="0.3.2"
-GridapGmsh="0.7.1"
