### 2025-01-15 53ff574
re-use sparsity pattern, and compute laplacian
