+++ src/Commons/SolveProblem.jl
@@ -36,9 +36,10 @@ function initialize_solve(simcase::SimulationCase,params::Dict{Symbol,Any})
 
 
   uh0, ph0 = create_initial_conditions(simcase,params)
+  ∇uh0 = create_initial_∇uh(uh0,params)
   @info "Initial Conditions Created"
 
-  matrices = initialize_matrices(uh0, params,simcase)
+  matrices = initialize_matrices((uh0,∇uh0), params,simcase)
   vectors = initialize_vectors(matrices,uh0,ph0)
 
 
@@ -47,8 +48,7 @@ function initialize_solve(simcase::SimulationCase,params::Dict{Symbol,Any})
   mkpath(save_sim_dir)
 
 
-  println("time 0")
-  @time uh_avg = FEFunction(Ut0, vectors[2])
+  uh_avg = FEFunction(Ut0, vectors[2])
   set_zeros!(uh_avg.fields)
   ph_avg = FEFunction(Pt0,  vectors[1])
   set_zeros!(ph_avg.fields)
@@ -85,13 +85,15 @@ GridapPETSc.with(args=split(petsc_options)) do
 
 
 Mat_Tuu, Mat_Tpu, Mat_Auu, Mat_Aup, Mat_Apu, Mat_App, 
-  Mat_ML, Mat_inv_ML, Mat_S, Vec_Au, Vec_Ap = matrices
+  Mat_ML, Mat_inv_ML, Mat_S, Vec_Auu, Vec_Aup, Vec_Apu, Vec_App, Vec_Au, Vec_Ap = matrices
 
 vec_pm,vec_um,vec_am,vec_sum_pm,Δa_star,Δpm1,Δa,b1,b2,ũ_vector = vectors
 
 ns1 = create_PETSc_setup(Mat_ML,vel_kspsetup)
 ns2 = create_PETSc_setup(Mat_S,pres_kspsetup)
+
 uh_tn_updt = FEFunction(Utn, vec_um)
+∇uh_tn = interpolate(∇(uh_tn_updt),params[:∇U])
 
 for (ntime,tn) in enumerate(time_step)
 
@@ -101,12 +103,7 @@ for (ntime,tn) in enumerate(time_step)
         println("update_matrices")
     
         @time begin
-         Mat_Tuu, Mat_Tpu, Mat_Auu, Mat_Aup, Mat_Apu, Mat_App, 
-          Mat_ML, Mat_inv_ML, Mat_S, Vec_Au, Vec_Ap = compute_matrices(uh_tn_updt, params, simcase)
-         
-          matrices = ( Mat_Tuu, Mat_Tpu, Mat_Auu, Mat_Aup, Mat_Apu, Mat_App, 
-          Mat_ML, Mat_inv_ML, Mat_S, Vec_Au, Vec_Ap)
-
+          update_all_matrices_vectors!(matrices, (uh_tn_updt,∇uh_tn), params,simcase)
         end
 
         println("update numerical set up")
@@ -143,22 +140,22 @@ for (ntime,tn) in enumerate(time_step)
         solve_pressure!(ns2,matrices,vectors,dt,θ)
       
      
-           Δpm1 = GridapDistributed.change_ghost(Δpm1, Mat_Aup)
+        Δpm1 = GridapDistributed.change_ghost(Δpm1, Mat_Aup)
+
         println("Update Δa")
         @time Δa .= Δa_star - θ .* Mat_inv_ML .* (Mat_Aup * Δpm1)
 
         vec_um .+=  dt * Δa
         vec_pm .+= Δpm1
 
[…]
