+++ src/Commons/CreateProblem/InitialConditions.jl
@@ -24,12 +24,20 @@ function create_initial_conditions(simcase::VelocityBoundaryCase, params::Dict{S
         uh0 = interpolate(uh_0, Ut0)
         ph0 = interpolate(ph_0, Pt0)
     end
-
+    
+    
     print_initial_conditions((uh0,ph0,uh0,uh0,ph0), simcase,params)
 
     return uh0,ph0
 end
 
+function create_initial_∇uh(uh0,params)
+    @unpack ∇U = params ## initial 
+
+    ∇uh =  interpolate(∇(uh0),∇U)
+
+    return ∇uh
+end
 
 #TaylorGreenCase Natural and Periodic
 function create_initial_conditions(simcase::TaylorGreen,params::Dict{Symbol,Any})

[…]
