### 2025-01-15 ef3432f
laplacian test
