+++ test/CommonsTests/CreateProblemTests/InitialConditionsTests.jl
@@ -29,8 +29,8 @@ function test_initialconditions(rank_partition, distribute, D)
     
         boundary_conditions = create_boundary_conditions(simcase) 
     
-        V, U, P, Q, Y, X = creation_fe_spaces(simcase, model, boundary_conditions)
-       
+        V, U, P, Q = creation_fe_spaces(simcase, model, boundary_conditions)
+        ∇U = creation_∇fe_space(simcase, model)
 
         trials = [U, P]
         tests = [V, Q]
@@ -43,10 +43,9 @@ function test_initialconditions(rank_partition, distribute, D)
         new_dict = Dict(:parts=>parts,
         :U => U,
         :P => P,
-        :X => X,
-        :Y => Y,
         :Ω => Ω,
         :dΩ => dΩ,
+        :∇U => ∇U,
         :degree => degree,
         :trials => trials, 
         :tests => tests)
