### 2025-01-15 #7564 @carlodev
@marinlauber thanks for your spot on comments, I will work on them, and on the issues open in the next few days.
I will also solve the two editorial issues
