### 2025-01-15 1e77941
compute laplacian
