+++ src/Commons/ExportUtility.jl
@@ -401,8 +401,6 @@ end
 
 function writesolution(simcase::VelocityBoundaryCase, Ω, nsub::Int64, save_path, tn, fields::Tuple)
     uh_tn, ph_tn, uh_tn_updt, uh_avg, ph_avg = fields
-    @info "Number of subcells $nsub"
-
     writevtk(Ω, save_path, nsubcells=nsub, cellfields=["uh" => uh_tn, "uh_updt" => uh_tn_updt, "ph" => ph_tn,
         "uh_avg" => uh_avg, "ph_avg" => ph_avg])
 end
