+++ test/CommonsTests/CreateProblemTests/InitialConditionsTests.jl
@@ -30,7 +30,6 @@ function test_initialconditions(rank_partition, distribute, D)
         boundary_conditions = create_boundary_conditions(simcase) 
     
         V, U, P, Q = creation_fe_spaces(simcase, model, boundary_conditions)
-        ∇U = creation_∇fe_space(simcase, model)
 
         trials = [U, P]
         tests = [V, Q]
@@ -45,7 +44,6 @@ function test_initialconditions(rank_partition, distribute, D)
         :P => P,
         :Ω => Ω,
         :dΩ => dΩ,
-        :∇U => ∇U,
         :degree => degree,
         :trials => trials, 
         :tests => tests)
