+++ docs/make.jl
@@ -17,6 +17,7 @@ makedocs(;plugins=[bib],
 
         "MPI run" => "mpi.md",
         "Examples" =>[
+        "Intro"=>"Cases/examples.md",
         "Taylor Green" => "Cases/taylorgreen.md",
         "Lid Driven Cavity Flow" => "Cases/liddriven.md",        
         "Cylinder" => "Cases/cylinder.md",
@@ -28,8 +29,9 @@ makedocs(;plugins=[bib],
             "FEM"=>"Theory/FEMTheory.md",
             "SUPG & VMS Stabilization"=>"Theory/SUPG_VMS_stab.md"],
         "Tools"=>[       
-        "Post Processing" => "post_proc.md",
-        "Boundary Layer Initialization" => "blinit.md"],
+        "Visualization" => "Tools/visualization.md",
+        "Post Processing" => "Tools/post_proc.md",
+        "Boundary Layer Initialization" => "Tools/blinit.md"],
         "API information" => "api_info.md",
         "References" => "references.md",
     ],
