+++ docs/src/Cases/examples.md
@@ -1,0 +1,5 @@
+# Examples
+
+In this section of the documentation you can find some examples explained. In the folder `/examples` of the repository you can find some examples to run in the `REPL`.
+
+To run it in parallel, give it a look at 
