+++ src/Commons/MatrixCreation.jl
@@ -98,9 +98,12 @@ end
 
 
 function initialize_matrices(u_adv, params,simcase)
+  @info "allocation matrix and vectors"
   matrices = allocate_all_matrices_vectors(u_adv, params,simcase)
   @info "matrix and vectors allocated"
-  update_all_matrices_vectors!(matrices,u_adv, params,simcase)
+
+  @info "updating matrix and vectors"
+  @time update_all_matrices_vectors!(matrices,u_adv, params,simcase)
   @info "matrix and vectors updated"
 
   return matrices
