+++ src/Commons/Equations/StabilizedEquations.jl
@@ -22,7 +22,7 @@ function segregated_equations(u_adv,params::Dict{Symbol,Any},simcase::Simulation
     Tc = continuity_stabilization(u_adv, stab_coeff, simcase)
   
     ### VMS extra equations
-    Auu_vms1(u, v) =  ∫(vms_activation .* u_adv ⋅ (∇(v))'*Tm⊙ (cconv ∘ (u_adv, ∇(u)) + visc_term ) )dΩ + 
+    Auu_vms1(u, v) =  ∫(vms_activation .* u_adv ⋅ (∇(v))'*Tm⊙ (cconv ∘ (u_adv, ∇(u)) ) )dΩ + 
      ∫((vms_activation *skewcoeff) .* u_adv ⋅ (∇(v))'*Tm⊙ (u_adv ⋅ (∇ ⋅ u)) )dΩ
 
     Tuu_vms1(u, v) =  ∫(vms_activation .* u_adv ⋅ (∇(v))'*Tm⊙u )dΩ # VMS1
@@ -32,7 +32,7 @@ function segregated_equations(u_adv,params::Dict{Symbol,Any},simcase::Simulation
     #### SUPG equations
     Tuu(u, v) = ∫((v + Tm * (cconv ∘ (u_adv, ∇(v)))) ⊙ u)dΩ + Tuu_vms1(u, v) 
     Tpu(u, q) = ∫(Tm * (∇(q)) ⊙ u)dΩ
-    Auu1(u, v) = ∫(ν * ∇(v) ⊙ ∇(u) + (cconv ∘ (u_adv, ∇(u))) ⋅ v + (Tm * (cconv ∘ (u_adv, ∇(v)))) ⊙ (cconv ∘ (u_adv, ∇(u)) + visc_term))dΩ
+    Auu1(u, v) = ∫(ν * ∇(v) ⊙ ∇(u) + (cconv ∘ (u_adv, ∇(u))) ⋅ v + (Tm * (cconv ∘ (u_adv, ∇(v)))) ⊙ (cconv ∘ (u_adv, ∇(u)) ))dΩ
   
     Auu2(u, v) = ∫((Tc * (∇ ⋅ v)) ⊙ (∇ ⋅ u) + skewcoeff .* u_adv ⋅ (v + Tm * (cconv ∘ (u_adv, ∇(v)))) ⋅ (∇ ⋅ u))dΩ
   
@@ -40,7 +40,7 @@ function segregated_equations(u_adv,params::Dict{Symbol,Any},simcase::Simulation
   
     Aup(p, v) = ∫( v⋅(∇(p) ) + (Tm * (cconv ∘ (u_adv, ∇(v)))) ⊙ ∇(p))dΩ + Aup_vms1(p, v)
   
-    Apu(u, q) = ∫(q * (∇ ⋅ u) + skewcoeff .* Tm ⋅ (∇(q)) ⋅ u_adv ⋅ (∇ ⋅ u) + Tm * (∇(q)) ⊙ (cconv ∘ (u_adv, ∇(u)) + visc_term))dΩ
+    Apu(u, q) = ∫(q * (∇ ⋅ u) + skewcoeff .* Tm ⋅ (∇(q)) ⋅ u_adv ⋅ (∇ ⋅ u) + Tm * (∇(q)) ⊙ (cconv ∘ (u_adv, ∇(u)) ))dΩ
   
     App(p, q) = ∫((Tm * ∇(q)) ⊙ (∇(p)))dΩ
   
