+++ Project.toml
@@ -34,10 +34,10 @@ UnPack = "3a884ed6-31ef-47d7-9d2a-63182c4928ed"
 [compat]
 DataInterpolations = "6.6.0"
 DelimitedFiles = "1.9.1"
-Gridap = "0.17.20"
-GridapDistributed = "0.3.2"
-GridapGmsh = "0.7.1"
-GridapPETSc = "0.5.3"
+Gridap = "=0.17.20"
+GridapDistributed = "=0.3.2"
+GridapGmsh = "=0.7.1"
+GridapPETSc = "=0.5.3"
 LinearAlgebra = "1.11.0"
 SparseArrays = "1.11.0"
 Test = "1.11.0"
