+++ README.md
@@ -12,9 +12,9 @@
 ![Julia flow](docs/figs/DU89.gif)
 
 ## Introduction
-The package solves the incompressible Navier-Stokes equations using SUPG and VMS method. VMS has been originally introduced by [Hughes2000](@cite). In specific, a linearized and segregated version of the SUPG (following the steps illustrated by [Janssens2014](@cite)) and VMS is solved. 
+The package solves the incompressible Navier-Stokes equations in the Finite Element Framework using SUPG and VMS method. VMS has been originally introduced by [Hughes2000](@cite). In specific, a linearized and segregated version of the SUPG (following the steps illustrated by [Janssens2014](@cite)) and VMS is solved. 
 
-The methods belong to the Large Eddy Simulation (LES) family. The package can solve the `Taylor Green Vortices` (only 2D), `Lid Driven Cavity Flow` (only 2D), `Cylinder vortex shedding` (only 2D), and general `Airfoils` (2D and 3D). 
+The methods belong to the Large Eddy Simulation (LES) family. The package can solve the `Taylor Green Vortices`, `Lid Driven Cavity Flow` (only 2D), `Cylinder vortex shedding` and and general `Airfoils` (2D and 3D). 
 
 It works fully in parallel (tested up to 80-CORES). It is specialized for the resolution of flow over airfoils, testing the capability of detecting the Laminar Separation Bubble. It is equipped with some utility `modules` for reading the output files and creating proper initial conditions.
 
