+++ test/CommonsTests/EquationsTests.jl
@@ -1,73 +1,0 @@
-module EquationsTests
-
-using Test
-using SegregatedVMSSolver
-using Gridap
-using Gridap.Fields
-using GridapDistributed
-using PartitionedArrays
-
-using SegregatedVMSSolver.ParametersDef
-using SegregatedVMSSolver.CreateProblem
-using SegregatedVMSSolver.Equations
- 
-
-
-include(joinpath("..","case_test.jl")) 
-
-
-function test_equations(rank_partition, distribute, D)
-    parts  = distribute(LinearIndices((prod(rank_partition),)))
-    
-    params = Dict{Symbol,Any}()
-
-    TestCase,mesh_file = first(iterate_test_cases(D))
-    simcase = create_simulation_test(TestCase, D; meshfile = mesh_file)
-    @sunpack order = simcase
-
-        model = create_model(parts, simcase)
-    
-        boundary_conditions = create_boundary_conditions(simcase) 
-    
-        V, U, P, Q  = creation_fe_spaces(simcase, model, boundary_conditions)
-
-        trials = [U, P]
-        tests = [V, Q]
-        
-        degree = 4*order
-        Ω = Triangulation(model)
-        dΩ = Measure(Ω, degree)
-    
-        
-        new_dict = Dict(:parts=>parts,
-        :U => U,
-        :P => P,
-        :Ω => Ω,
-        :dΩ => dΩ,
-        :degree => degree,
-        :trials => trials, 
-        :tests => tests)
-        merge!(params, new_dict)
-        
-        u_adv = interpolate_everywhere(VectorValue(zeros(D)...), U(0.0))
-
-
-        Tuu,Tpu,Auu,Aup,Apu,App,ML,S,rhs =segregated_equations(u_adv,params,simcase)
-        return true
-
-end
-
-
-
-function main(distribute)
-        D = 2
-        rank_partition = (2, 2)
-
-        @test test_equations(rank_partition, distribute, D)
-end
-
-
-
-
-
-end
