+++ test/CommonsTests/TurbParamsTests.jl
@@ -12,7 +12,7 @@ using SegregatedVMSSolver.ParametersDef
 function main(distribute)
     TI = 0.001    
     D = 3
-    meshfile = joinpath(@__DIR__, "..", "models", "sd7003s_3D_simple.msh")
+    meshfile = joinpath(@__DIR__, "..","..", "models", "sd7003s_3D_simple.msh")
     rank_partition=(2,2,1)
 
     #VirtualBox creation
@@ -43,7 +43,7 @@ function main(distribute)
     Vboxinfo.σ
     
     for ep in eval_point
-        @time compute_fluctuation(ep, tt, (TurbulenceInlet,Eddies, u_in_mag, Vboxinfo, Re_stress,D))
+        compute_fluctuation(ep, tt, (TurbulenceInlet,Eddies, 1.0, Vboxinfo, Re_stress,D,1.0))
     end
     
     
