+++ src/Commons/MatrixCreation.jl
@@ -24,7 +24,7 @@ export update_all_matrices_vectors!
 
 It allocates a zero vector where to store the inverse of the lumped matrix
 """
-function allocate_Mat_inv_ML(Mat_ML::PSparseMatrix) 
+function allocate_Mat_inv_ML(Mat_ML::PSparseMatrix)
   return pzeros(Mat_ML.row_partition)
 end
 
@@ -42,22 +42,22 @@ end
 
 It computes the lumped matrix, takes the inverse of the diagonal elements.
 """
-function inv_lump_vel_mass!(Mat_inv_ML::PVector,Mat_ML::PSparseMatrix)
-    values = map(Mat_ML.matrix_partition) do val
-        N = maximum(rowvals(val))
-    
-        V = zeros(N)
-        vals = nonzeros(val)
-        
-        j = 1
-        for i in rowvals(val)
-            V[i] += vals[j]
-            j+1
-        end
-        V = 1 ./V
+function inv_lump_vel_mass!(Mat_inv_ML::PVector, Mat_ML::PSparseMatrix)
+  values = map(Mat_ML.matrix_partition) do val
+    N = maximum(rowvals(val))
 
+    V = zeros(N)
+    vals = nonzeros(val)
+
+    j = 1
+    for i in rowvals(val)
+      V[i] += vals[j]
+      j + 1
     end
-    Mat_inv_ML .= PVector(values,Mat_ML.row_partition) 
+    V = 1 ./ V
+
+  end
+  Mat_inv_ML .= PVector(values, Mat_ML.row_partition)
 end
 
 
@@ -66,9 +66,9 @@ end
 #       if !isempty(inv_ML_vec[inv_ML_vec.==Inf])
 #       error("The matrix ML can not be inverted because after lumping zero values are detected")
 #   end
-  
+
 #   Mat_inv_ML.=inv_ML_vec
-  
+
 # end
 
 """
@@ -76,7 +76,7 @@ end
 
 It initializes vectors where velocity, pressure, acceleration and all the increments will be stored.
 """
-function initialize_vectors(matrices::Tuple,uh0,ph0)
+function initialize_vectors(matrices::Tuple, uh0, ph0)
   Mat_Tuu, Mat_Tpu, Mat_Auu, Mat_Aup, Mat_Apu, Mat_App, Mat_ML, Mat_inv_ML, Mat_S, Vec_Auu, Vec_Aup, Vec_Apu, Vec_App, Vec_Au, Vec_Ap = matrices
   vec_pm = GridapDistributed.change_ghost(get_free_dof_values(ph0), Mat_Aup)
   vec_um = GridapDistributed.change_ghost(get_free_dof_values(uh0), Mat_Auu)
@@ -92,72 +92,74 @@ function initialize_vectors(matrices::Tuple,uh0,ph0)
   b2 = pazeros(Vec_Ap)
   ũ_vector = create_ũ_vector(vec_um)
 
-  return vec_pm,vec_um,vec_am,vec_sum_pm,Δa_star,Δpm1,Δa,b1,b2,ũ_vector
+  return vec_pm, vec_um, vec_am, vec_sum_pm, Δa_star, Δpm1, Δa, b1, b2, ũ_vector
 end
 
 
 
-function initialize_matrices(u_adv, params,simcase)
[…]
