+++ .github/workflows/CI.yml
@@ -6,8 +6,8 @@ on:
       - refactor-matrix-assemble
     tags: ['*']
 jobs:
-  VMSTests:
-    name: VMS Tests
+  REPLTests:
+    name: SegregatedVMSSolver-Sequential
     runs-on: ${{ matrix.os }}
     strategy:
       fail-fast: false
@@ -35,10 +35,41 @@ jobs:
             ${{ runner.os }}-test-
             ${{ runner.os }}-
       - uses: julia-actions/julia-buildpkg@v1
-      - uses: julia-actions/julia-runtest@v1
+      - run: julia --project=. -e "include(joinpath(\"test\", \"sequential\", \"runtests.jl\"))"
       - uses: julia-actions/julia-processcoverage@v1
       - uses: codecov/codecov-action@v4
         env:
           CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
         with:
-          threshold:  70%
+          threshold:  50%
+
+  MPITests:
+    name: SegregatedVMSSolver-MPI
+    runs-on: ${{ matrix.os }}
+    strategy:
+      fail-fast: false
+      matrix:
+        version:
+          - '1.11.1'
+        os:
+          - ubuntu-latest
+        arch:
+          - x64
+    steps:
+      - uses: actions/checkout@v3
+      - uses: julia-actions/setup-julia@latest
+        with:
+          version: ${{ matrix.version }}
+          arch: ${{ matrix.arch }}
+      - uses: actions/cache@v1
+        env:
+          cache-name: cache-artifacts
+        with:
+          path: ~/.julia/artifacts
+          key: ${{ runner.os }}-mpi-test-${{ env.cache-name }}-${{ hashFiles('**/Project.toml') }}
+          restore-keys: |
+            ${{ runner.os }}-mpi-test-${{ env.cache-name }}-
+            ${{ runner.os }}-mpi-test-
+            ${{ runner.os }}-
+      - uses: julia-actions/julia-buildpkg@v1
+      - run: julia --project=. -e "include(joinpath(\"test\", \"mpi\", \"runtests.jl\"))"
