### 2025-01-22 bbc7754
rm extra prints
