+++ versions.js
@@ -1,8 +1,9 @@
 var DOC_VERSIONS = [
   "stable",
+  "v3.0",
   "v2.0",
   "v1.0",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v2.0.1";
+var DOCUMENTER_NEWEST = "v3.0.0";
 var DOCUMENTER_STABLE = "stable";
