+++ paper/paper.bib
@@ -38,7 +38,6 @@
 }
 
 @article{Calderer:2013,
-   abstract = {This paper presents three-level residual-based turbulence models for the incompressible Navier-Stokes equations. Employing the variational multiscale (VMS) framework, the velocity and pressure fields are decomposed into two overlapping hierarchical scales, thereby leading to a system of coupled mixed field problems. The mixed problem at the fine scales is stabilized via a subsequent VMS application that results in a further decomposition of the fine-scale velocity field into level-I and level-II scales. The level-II scales are modeled using higher-order bubble functions that are then variationally embedded in the level-I formulation to stabilize it. The level-I problem is modeled via a second set of bubble functions that are linearly independent of the bubbles employed at level-II. Finally, the resulting level-I fine-scales are variationally embedded in the coarse-scale formulation. This yields a residual-based turbulence model for the larger or coarser-scales. A significant feature of the proposed method is that it results in a concurrent top-down and bottom up two-way nesting of the scales. In addition, the resulting turbulence model does not possess any embedded tunable parameters. Another attribute of the formulation is that the fine scales at every level are driven by the residuals of the Euler-Lagrange equations of the coarser scales at the preceding levels, thereby resulting in a method that is variationally consistent. Various algorithmic generalizations of the method are presented that lead to computationally economic residual-based turbulence models. The proposed telescopic depth in scales approach helps make these models accurate for low order tetrahedral and hexahedral elements, a feature that is facilitated by the higher-order bubble functions over element interiors and it results in an enhanced representation of the fine-scale terms modeling the fine viscous effects. From a computational perspective this method results in easy-to-implement equal-order pressure-velocity elements, and possesses the desirable p-refinement feature. Numerical performance of the method is assessed on turbulent channel flow problems at Re = 395 and Re = 590. Also presented are the results for turbulent SD-7003 airfoil at Re = 60, 000 and comparison is made with the published experimental data and numerical results. © 2012 Elsevier B.V.},
    author = {Ramon Calderer and Arif Masud},
    doi = {10.1016/j.cma.2012.09.015},
    issn = {00457825},
@@ -51,22 +50,25 @@
    year = {2013},
 }
 
-@misc{Galbraith:2008,
[…]
