+++ .github/workflows/CI.yml
@@ -3,7 +3,6 @@ on:
   push:
     branches:
       - master
-      - refactor-matrix-assemble
     tags: ['*']
 jobs:
   REPLTests:
