+++ test/runtests.jl
@@ -1,0 +1,12 @@
+module SegregatedVMSSolverTests
+
+using Test
+
+
+@testset "Sequential" begin  include("sequential/runtests.jl") end
+
+@testset "Utils" begin  include("UtilsTests/runtests.jl") end
+
+
+
+end
