### 2025-01-23 carlodev/SegregatedVMSSolver.jl#25 @carlodev
added in readme.txt at the end the `Contributing` section
