### 2025-01-23 carlodev/SegregatedVMSSolver.jl#26 @carlodev
`examples/` folder created, with `2D` and `3D` examples that can be run
