### 2025-01-23 carlodev/SegregatedVMSSolver.jl#24 @carlodev
New version released (v3.0.0), they have been fixed
