### 2025-01-23 carlodev/SegregatedVMSSolver.jl#27 @carlodev
New version v3.0.0 released, it should be fixed
