### 2025-01-23 c409711
add journal paper
