+++ paper/paper.bib
@@ -51,9 +51,10 @@
 }
 
 @article{@Galbraith:2008,
-  title = {Implicit {{Large Eddy Simulation}} of {{Low Reynolds Number Flow Past}} the {{SD7003 Airfoil}}},
+  title = {Implicit Large Eddy Simulation of Low Reynolds Number Flow Past the SD7003 Airfoil},
   author = {Marshall C Galbraith and Miguel R Visbal},
   year = {2008},
+  journal = {40th Fluid Dynamics Conference and Exhibit},
   doi = {10.2514/6.2010-4737},
   keywords = {les,sd7003},
 }
