+++ paper/paper.bib
@@ -50,7 +50,7 @@
    year = {2013},
 }
 
-@article{@Galbraith:2008,
+@article{Galbraith:2008,
   title = {Implicit Large Eddy Simulation of Low Reynolds Number Flow Past the SD7003 Airfoil},
   author = {Marshall C Galbraith and Miguel R Visbal},
   year = {2008},
