### 2025-01-23 #7564 @carlodev
@Kevin-Mattheus-Moerman I fixed the editorial issues you pointed out

The issues opened by @marinlauber @nathanquinlan have been answered
The paper and the documentation have been upgraded taking into account the remarks of @marinlauber

I remain at your disposal for any new issue/suggestion
