### 2025-01-24 carlodev/SegregatedVMSSolver.jl#26 @marinlauber
Thanks for providing that. The last comment before we can close this issue is that the file `examples/3D/Airfoil3D.jl` is empty.
