### 2025-01-24 carlodev/SegregatedVMSSolver.jl#24 @marinlauber
With the clone of the github version, I still get an error

```
(SegregatedVMSSolver) pkg> test
     Testing SegregatedVMSSolver
      Status `/tmp/jl_yl2Iwj/Project.toml`
  [336ed68f] CSV v0.10.15
  [a93c6f00] DataFrames v1.7.0
⌅ [82cc6244] DataInterpolations v6.6.0
  [8bb1440f] DelimitedFiles v1.9.1
  [7a1cc6ca] FFTW v1.8.1
  [5789e2e9] FileIO v1.16.6
  [1a297f60] FillArrays v1.13.0
⌅ [56d4f2e9] Gridap v0.17.20
⌅ [f9701e48] GridapDistributed v0.3.2
⌅ [3025c34a] GridapGmsh v0.7.1
  [bcdc36c2] GridapPETSc v0.5.3
  [da04e1cc] MPI v0.20.22
  [b8a86587] NearestNeighbors v0.4.21
  [d96e819e] Parameters v0.12.3
⌅ [5a9dfac6] PartitionedArrays v0.3.4
  [295af30f] Revise v3.7.2
  [3f865c0f] ScatteredInterpolation v0.3.6
  [90220dfc] SegregatedVMSSolver v3.0.0 `~/Workspace/SegregatedVMSSolver.jl`
  [10745b16] Statistics v1.11.1
  [57cf2c78] SyntheticEddyMethod v0.4.4
  [592b5752] Trapz v2.0.3
  [3a884ed6] UnPack v1.0.2
  [37e2e46d] LinearAlgebra v1.11.0
  [44cfe95a] Pkg v1.11.0
  [9a3f8284] Random v1.11.0
  [2f01184e] SparseArrays v1.11.0
  [8dfed614] Test v1.11.0
      Status `/tmp/jl_yl2Iwj/Manifest.toml`
  [621f4979] AbstractFFTs v1.5.0
  [1520ce14] AbstractTrees v0.4.5
  [79e6a3ab] Adapt v4.1.1
  [dce04be8] ArgCheck v2.4.0
  [4fba245c] ArrayInterface v7.18.0
  [4c555306] ArrayLayouts v1.11.0
  [15f4f7f2] AutoHashEquals v2.2.0
  [13072b0f] AxisAlgorithms v1.1.0
  [fbb218c0] BSON v0.3.9
⌅ [8e7c35d0] BlockArrays v0.16.43
  [fa961155] CEnum v0.5.0
  [336ed68f] CSV v0.10.15
  [d360d2e6] ChainRulesCore v1.25.1
  [7a955b69] CircularArrays v1.4.0
  [da1fd8a2] CodeTracking v1.3.6
  [6309b1aa] CodecInflate64 v0.1.3
  [944b1d66] CodecZlib v0.7.6
  [861a8166] Combinatorics v1.0.2
  [bbf7d656] CommonSubexpressions v0.3.1
  [34da2185] Compat v4.16.0
  [187b0558] ConstructionBase v1.5.8
  [a8cc5b0e] Crayons v4.1.1
  [9a962f9c] DataAPI v1.16.0
  [a93c6f00] DataFrames v1.7.0
⌅ [82cc6244] DataInterpolations v6.6.0
  [864edb3b] DataStructures v0.18.20
  [e2d170a0] DataValueInterfaces v1.0.0
  [8bb1440f] DelimitedFiles v1.9.1
  [163ba53b] DiffResults v1.1.0
  [b552c78f] DiffRules v1.15.1
  [b4f34e82] Distances v0.10.12
  [ffbed154] DocStringExtensions v0.9.3
  [8f5d6c58] EzXML v1.2.0
  [7a1cc6ca] FFTW v1.8.1
⌅ [442a2c76] FastGaussQuadrature v0.4.9
  [5789e2e9] FileIO v1.16.6
  [48062228] FilePathsBase v0.9.22
  [1a297f60] FillArrays v1.13.0
  [64ca27bc] FindFirstFunctions v1.4.1
  [6a86dc24] FiniteDiff v2.26.2
  [f6369f11] ForwardDiff v0.10.38
⌅ [56d4f2e9] Gridap v0.17.20
⌅ [f9701e48] GridapDistributed v0.3.2
⌅ [3025c34a] GridapGmsh v0.7.1
  [bcdc36c2] GridapPETSc v0.5.3
  [842dd82b] InlineStrings v1.4.2
  [0c81fc1b] InputBuffers v1.0.0
  [a98d9a8b] Interpolations v0.15.1
  [41ab1584] InvertedIndices v1.3.1
  [92d709cd] IrrationalConstants v0.2.2
  [42fd0dbc] IterativeSolvers v0.9.4
  [82899510] IteratorInterfaceExtensions v1.0.0
⌅ [033835bb] JLD2 v0.4.53
  [692b3bcd] JLLWrappers v1.7.0
  [682c06a0] JSON v0.21.4
  [aa1ae85d] JuliaInterpreter v0.9.39
  [b964fa9f] LaTeXStrings v1.4.0
  [9c8b4983] LightXML v0.9.1
  [d3d80556] LineSearches v7.3.0
  [2ab3a3ac] LogExpFunctions v0.3.29
  [6f1432cf] LoweredCodeUtils v3.1.0
  [da04e1cc] MPI v0.20.22
  [3da0fdf6] MPIPreferences v0.1.11
  [1914dd2f] MacroTools v0.5.15
  [2679e427] Metis v1.5.0
  [e1d29d7a] Missings v1.2.0
  [d41bc354] NLSolversBase v7.8.3
  [2774e3e8] NLsolve v4.5.1
  [77ba4419] NaNMath v1.1.1
  [b8a86587] NearestNeighbors v0.4.21
  [6fe1bfb0] OffsetArrays v1.15.0
  [bac558e1] OrderedCollections v1.7.0
  [d96e819e] Parameters v0.12.3
  [69de0a69] Parsers v2.8.1
⌅ [5a9dfac6] PartitionedArrays v0.3.4
  [eebad327] PkgVersion v0.3.3
  [c74db56a] PolynomialBases v0.4.22
  [2dfb63ee] PooledArrays v1.4.3
  [aea7be01] PrecompileTools v1.2.1
  [21216c6a] Preferences v1.4.3
  [08abe8d2] PrettyTables v2.4.0
  [1fd47b50] QuadGK v2.11.1
  [c84ed2f1] Ratios v0.4.5
  [3cdcf5f2] RecipesBase v1.3.4
  [189a3867] Reexport v1.2.2
  [ae029012] Requires v1.3.0
  [295af30f] Revise v3.7.2
  [3f865c0f] ScatteredInterpolation v0.3.6
  [90220dfc] SegregatedVMSSolver v3.0.0 `~/Workspace/SegregatedVMSSolver.jl`
  [91c51154] SentinelArrays v1.4.8
  [efcf1570] Setfield v1.1.1
  [ce78b400] SimpleUnPack v1.1.0
  [a2af1166] SortingAlgorithms v1.2.1
  [a0a7dd2c] SparseMatricesCSR v0.6.8
  [276daf66] SpecialFunctions v2.5.0
  [90137ffa] StaticArrays v1.9.10
  [1e83bf80] StaticArraysCore v1.4.3
  [10745b16] Statistics v1.11.1
  [82ae8749] StatsAPI v1.7.0
  [892a3eda] StringManipulation v0.4.0
  [57cf2c78] SyntheticEddyMethod v0.4.4
  [3783bdb8] TableTraits v1.0.1
  [bd369af6] Tables v1.12.0
  [3bb67fe8] TranscodingStreams v0.11.3
  [592b5752] Trapz v2.0.3
  [3a884ed6] UnPack v1.0.2
  [4004b06d] VTKBase v1.0.1
  [ea10d353] WeakRefStrings v1.4.2
  [efce3f68] WoodburyMatrices v1.0.0
  [76eceee3] WorkerUtilities v1.6.1
  [64499a7a] WriteVTK v1.21.1
  [fdbf4ff8] XLSX v0.10.4
  [49080126] ZipArchives v2.4.0
  [a5390f91] ZipFile v0.10.1
  [6e34b625] Bzip2_jll v1.0.8+4
  [83423d85] Cairo_jll v1.18.2+1
  [2e619515] Expat_jll v2.6.4+3
  [f5851436] FFTW_jll v3.3.10+3
  [4fce6fc7] FLTK_jll v1.3.8+0
  [a3f928ae] Fontconfig_jll v2.15.0+0
  [d7e528f0] FreeType2_jll v2.13.3+1
  [bd17208b] GLU_jll v9.0.1+0
  [78b55507] Gettext_jll v0.21.0+0
  [7746bdde] Glib_jll v2.82.4+0
⌅ [0234f1f7] HDF5_jll v1.12.2+2
  [e33a78d0] Hwloc_jll v2.11.2+3
  [1d5cc7b8] IntelOpenMP_jll v2025.0.4+0
  [aacddb02] JpegTurbo_jll v3.1.1+0
  [1d63c593] LLVMOpenMP_jll v18.1.7+0
  [dd4b983a] LZO_jll v2.10.3+0
⌅ [e9f186c6] Libffi_jll v3.2.2+2
  [d4300ac3] Libgcrypt_jll v1.11.0+0
  [7e76a0d4] Libglvnd_jll v1.7.0+0
  [7add5ba3] Libgpg_error_jll v1.51.1+0
  [94ce4f54] Libiconv_jll v1.18.0+0
  [4b2f31a3] Libmount_jll v2.40.3+0
  [38a345b3] Libuuid_jll v2.40.3+0
  [18c40d15] LinearElasticity_jll v5.0.0+0
  [d00139f3] METIS_jll v5.1.3+0
  [856f044c] MKL_jll v2025.0.1+1
  [86086c02] MMG_jll v5.6.0+0
  [7cb0a576] MPICH_jll v4.2.3+0
  [f1f71cc9] MPItrampoline_jll v5.5.1+2
  [9237b28f] MicrosoftMPI_jll v10.1.4+3
⌅ [baad4e97] OCCT_jll v7.6.2+2
  [656ef2d0] OpenBLAS32_jll v0.3.29+0
  [fe0851c0] OpenMPI_jll v5.0.6+0
⌅ [458c3c95] OpenSSL_jll v1.1.23+1
  [efe28fd5] OpenSpecFun_jll v0.5.6+0
⌅ [8fa3689e] PETSc_jll v3.15.2+0
⌅ [30392449] Pixman_jll v0.43.4+0
⌅ [a8d0f55d] SCOTCH_jll v6.1.3+0
  [02c8fc9c] XML2_jll v2.13.5+0
  [aed1982a] XSLT_jll v1.1.42+0
  [4f6342f7] Xorg_libX11_jll v1.8.6+3
  [0c0b7dd1] Xorg_libXau_jll v1.0.12+0
  [a3789734] Xorg_libXdmcp_jll v1.1.5+0
  [1082639a] Xorg_libXext_jll v1.3.6+3
  [d091e8ba] Xorg_libXfixes_jll v6.0.0+0
  [2c808117] Xorg_libXft_jll v2.3.3+1
  [d1454406] Xorg_libXinerama_jll v1.1.5+0
  [ea2f1a96] Xorg_libXrender_jll v0.9.11+1
  [14d82f49] Xorg_libpthread_stubs_jll v0.1.2+0
  [c7cfdc94] Xorg_libxcb_jll v1.17.0+3
  [c5fb5394] Xorg_xtrans_jll v1.5.1+0
⌃ [630162c2] gmsh_jll v4.12.2+0
  [b53b4c65] libpng_jll v1.6.45+1
  [1317d2d5] oneTBB_jll v2021.12.0+0
  [0dad84c5] ArgTools v1.1.2
  [56f22d72] Artifacts v1.11.0
  [2a0f44e3] Base64 v1.11.0
  [ade2ca70] Dates v1.11.0
  [8ba89e20] Distributed v1.11.0
  [f43a241f] Downloads v1.6.0
  [7b1f6079] FileWatching v1.11.0
  [9fa8497b] Future v1.11.0
  [b77e0a4c] InteractiveUtils v1.11.0
  [4af54fe1] LazyArtifacts v1.11.0
  [b27032c2] LibCURL v0.6.4
  [76f85450] LibGit2 v1.11.0
  [8f399da3] Libdl v1.11.0
  [37e2e46d] LinearAlgebra v1.11.0
  [56ddb016] Logging v1.11.0
  [d6f4376e] Markdown v1.11.0
  [a63ad114] Mmap v1.11.0
  [ca575930] NetworkOptions v1.2.0
  [44cfe95a] Pkg v1.11.0
  [de0858da] Printf v1.11.0
  [3fa0cd96] REPL v1.11.0
  [9a3f8284] Random v1.11.0
  [ea8e919c] SHA v0.7.0
  [9e88b42a] Serialization v1.11.0
  [1a1011a3] SharedArrays v1.11.0
  [6462fe0b] Sockets v1.11.0
  [2f01184e] SparseArrays v1.11.0
  [f489334b] StyledStrings v1.11.0
  [4607b0f0] SuiteSparse
  [fa267f1f] TOML v1.0.3
  [a4e569a6] Tar v1.10.0
  [8dfed614] Test v1.11.0
  [cf7118a7] UUIDs v1.11.0
  [4ec0a83e] Unicode v1.11.0
  [e66e0078] CompilerSupportLibraries_jll v1.1.1+0
  [781609d7] GMP_jll v6.3.0+0
  [deac9b47] LibCURL_jll v8.6.0+0
  [e37daf67] LibGit2_jll v1.7.2+0
  [29816b5a] LibSSH2_jll v1.11.0+1
  [c8ffd9c3] MbedTLS_jll v2.28.6+0
  [14a3606d] MozillaCACerts_jll v2023.12.12
  [4536629a] OpenBLAS_jll v0.3.27+1
  [05823500] OpenLibm_jll v0.8.1+2
  [efcefdf7] PCRE2_jll v10.42.0+1
  [bea87d4a] SuiteSparse_jll v7.7.0+0
  [83775a58] Zlib_jll v1.2.13+1
  [8e850b90] libblastrampoline_jll v5.11.0+0
  [8e850ede] nghttp2_jll v1.59.0+0
  [3f19e933] p7zip_jll v17.4.0+2
        Info Packages marked with ⌃ and ⌅ have new versions available. Those with ⌃ may be upgradable, but those with ⌅ are restricted by compatibility constraints from upgrading.
     Testing Running tests...
WARNING: Method definition numerical_setup!(GridapPETSc.PETScLinearSolverNS{T} where T, AbstractArray{T, 2} where T) in module GridapPETSc at /home/marin/.julia/packages/GridapPETSc/X0jXh/src/PETScLinearSolvers.jl:101 overwritten in module SolverOptions at /home/marin/Workspace/SegregatedVMSSolver.jl/src/Commons/SolverOptions.jl:66.
WARNING: Method definition solve!(GridapPETSc.PETScVector, GridapPETSc.PETScLinearSolverNS{T} where T, AbstractArray{T, 1} where T) in module GridapPETSc at /home/marin/.julia/packages/GridapPETSc/X0jXh/src/PETScLinearSolvers.jl:67 overwritten in module SolverOptions at /home/marin/Workspace/SegregatedVMSSolver.jl/src/Commons/SolverOptions.jl:76.
New tag centre added to the model
[ Info: model read completed
[ Info: boundary conditions created
[ Info: FE Spaces Created
New tag centre added to the model
Info    : Reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/CommonsTests/CreateProblemTests/../../../models/DU89_2D_A1_M.msh'...
Info    : 500 entities
Info    : 38274 nodes
Info    : 38734 elements
Info    : Done reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/CommonsTests/CreateProblemTests/../../../models/DU89_2D_A1_M.msh'
[ Info: Restarting uh0 ...
[ Info: Restarting ph0 ...
Info    : Reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/../models/DU89_2D_A1_M.msh'...
Info    : 500 entities
Info    : 38274 nodes
Info    : 38734 elements
Info    : Done reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/../models/DU89_2D_A1_M.msh'
Info    : Reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/../models/Cylinder_2D.msh'...
Info    : 65 entities
Info    : 10839 nodes
Info    : 11054 elements
Info    : Done reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/../models/Cylinder_2D.msh'
New tag centre added to the model
Info    : Reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/../models/sd7003s_3D_simple.msh'...
Info    : 214 entities
Info    : 19554 nodes
Info    : 20867 elements
Info    : Done reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/../models/sd7003s_3D_simple.msh'
Info    : Reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/../models/DU89_2D_A1_M.msh'...
Info    : 500 entities
Info    : 38274 nodes
Info    : 38734 elements
Info    : Done reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/../models/DU89_2D_A1_M.msh'
Info    : Reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/../models/Cylinder_2D.msh'...
Info    : 65 entities
Info    : 10839 nodes
Info    : 11054 elements
Info    : Done reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/../models/Cylinder_2D.msh'
New tag centre added to the model
Info    : Reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/../models/sd7003s_3D_simple.msh'...
Info    : 214 entities
Info    : 19554 nodes
Info    : 20867 elements
Info    : Done reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/../models/sd7003s_3D_simple.msh'
Info    : Reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/../models/DU89_2D_A1_M.msh'...
Info    : 500 entities
Info    : 38274 nodes
Info    : 38734 elements
Info    : Done reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/../models/DU89_2D_A1_M.msh'
┌ Warning: detected extension '.vtu_1' does not correspond to dataset type.
│ Appending '.vtu' to filename.
└ @ WriteVTK ~/.julia/packages/WriteVTK/H5smu/src/WriteVTK.jl:162
┌ Warning: detected extension '.vtu_2' does not correspond to dataset type.
│ Appending '.vtu' to filename.
└ @ WriteVTK ~/.julia/packages/WriteVTK/H5smu/src/WriteVTK.jl:162
┌ Warning: detected extension '.vtu_3' does not correspond to dataset type.
│ Appending '.vtu' to filename.
└ @ WriteVTK ~/.julia/packages/WriteVTK/H5smu/src/WriteVTK.jl:162
┌ Warning: detected extension '.vtu_4' does not correspond to dataset type.
│ Appending '.vtu' to filename.
└ @ WriteVTK ~/.julia/packages/WriteVTK/H5smu/src/WriteVTK.jl:162
[ Info: allocation matrix and vectors
[ Info: matrix and vectors allocated
[ Info: updating matrix and vectors
  6.803545 seconds (11.62 M allocations: 664.240 MiB, 1.28% gc time, 65.36% compilation time)
[ Info: matrix and vectors updated
[ Info: Eddies initialized - total eddies: 46656
New tag centre added to the model
[ Info: model read completed
[ Info: boundary conditions created
[ Info: FE Spaces Created
New tag centre added to the model
Info    : Reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/CommonsTests/CreateProblemTests/../../../models/DU89_2D_A1_M.msh'...
Info    : 500 entities
Info    : 38274 nodes
Info    : 38734 elements
Info    : Done reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/CommonsTests/CreateProblemTests/../../../models/DU89_2D_A1_M.msh'
[ Info: Restarting uh0 ...
[ Info: Restarting ph0 ...
Info    : Reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/../models/DU89_2D_A1_M.msh'...
Info    : 500 entities
Info    : 38274 nodes
Info    : 38734 elements
Info    : Done reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/../models/DU89_2D_A1_M.msh'
Info    : Reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/../models/Cylinder_2D.msh'...
Info    : 65 entities
Info    : 10839 nodes
Info    : 11054 elements
Info    : Done reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/../models/Cylinder_2D.msh'
New tag centre added to the model
Info    : Reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/../models/sd7003s_3D_simple.msh'...
Info    : 214 entities
Info    : 19554 nodes
Info    : 20867 elements
Info    : Done reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/../models/sd7003s_3D_simple.msh'
Info    : Reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/../models/DU89_2D_A1_M.msh'...
Info    : 500 entities
Info    : 38274 nodes
Info    : 38734 elements
Info    : Done reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/../models/DU89_2D_A1_M.msh'
Info    : Reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/../models/Cylinder_2D.msh'...
Info    : 65 entities
Info    : 10839 nodes
Info    : 11054 elements
Info    : Done reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/../models/Cylinder_2D.msh'
New tag centre added to the model
Info    : Reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/../models/sd7003s_3D_simple.msh'...
Info    : 214 entities
Info    : 19554 nodes
Info    : 20867 elements
Info    : Done reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/../models/sd7003s_3D_simple.msh'
Info    : Reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/../models/DU89_2D_A1_M.msh'...
Info    : 500 entities
Info    : 38274 nodes
Info    : 38734 elements
Info    : Done reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/../models/DU89_2D_A1_M.msh'
┌ Warning: detected extension '.vtu_1' does not correspond to dataset type.
│ Appending '.vtu' to filename.
└ @ WriteVTK ~/.julia/packages/WriteVTK/H5smu/src/WriteVTK.jl:162
┌ Warning: detected extension '.vtu_2' does not correspond to dataset type.
│ Appending '.vtu' to filename.
└ @ WriteVTK ~/.julia/packages/WriteVTK/H5smu/src/WriteVTK.jl:162
┌ Warning: detected extension '.vtu_3' does not correspond to dataset type.
│ Appending '.vtu' to filename.
└ @ WriteVTK ~/.julia/packages/WriteVTK/H5smu/src/WriteVTK.jl:162
┌ Warning: detected extension '.vtu_4' does not correspond to dataset type.
│ Appending '.vtu' to filename.
└ @ WriteVTK ~/.julia/packages/WriteVTK/H5smu/src/WriteVTK.jl:162
[ Info: allocation matrix and vectors
[ Info: matrix and vectors allocated
[ Info: updating matrix and vectors
  2.578054 seconds (1.46 M allocations: 153.332 MiB, 0.76% gc time)
[ Info: matrix and vectors updated
[ Info: Eddies initialized - total eddies: 46656
[ Info: Eddies initialized - total eddies: 46656
Info    : Reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/TestsCases/../../models/DU89_2D_A1_M.msh'...
Info    : 500 entities
Info    : 38274 nodes
Info    : 38734 elements
Info    : Done reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/TestsCases/../../models/DU89_2D_A1_M.msh'
[ Info: model read completed
[ Info: boundary conditions created
[ Info: FE Spaces Created
Info    : Reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/TestsCases/../../models/DU89_2D_A1_M.msh'...
Info    : 500 entities
Info    : 38274 nodes
Info    : 38734 elements
Info    : Done reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/TestsCases/../../models/DU89_2D_A1_M.msh'
[ Info: model read completed
[ Info: boundary conditions created
[ Info: FE Spaces Created
Info    : Reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/TestsCases/../../models/Cylinder_2D.msh'...
Info    : 65 entities
Info    : 10839 nodes
Info    : 11054 elements
Info    : Done reading '/home/marin/Workspace/SegregatedVMSSolver.jl/test/TestsCases/../../models/Cylinder_2D.msh'
[ Info: model read completed
[ Info: boundary conditions created
[ Info: FE Spaces Created
┌ Warning: detected extension '.LidDriven_0_1' does not correspond to dataset type.
│ Appending '.vtu' to filename.
└ @ WriteVTK ~/.julia/packages/WriteVTK/H5smu/src/WriteVTK.jl:162
┌ Warning: detected extension '.LidDriven_0_2' does not correspond to dataset type.
│ Appending '.vtu' to filename.
└ @ WriteVTK ~/.julia/packages/WriteVTK/H5smu/src/WriteVTK.jl:162
┌ Warning: detected extension '.LidDriven_0_3' does not correspond to dataset type.
│ Appending '.vtu' to filename.
└ @ WriteVTK ~/.julia/packages/WriteVTK/H5smu/src/WriteVTK.jl:162
┌ Warning: detected extension '.LidDriven_0_4' does not correspond to dataset type.
│ Appending '.vtu' to filename.
└ @ WriteVTK ~/.julia/packages/WriteVTK/H5smu/src/WriteVTK.jl:162
┌ Warning: detected extension '.LidDriven_1_1' does not correspond to dataset type.
│ Appending '.vtu' to filename.
└ @ WriteVTK ~/.julia/packages/WriteVTK/H5smu/src/WriteVTK.jl:162
┌ Warning: detected extension '.LidDriven_1_2' does not correspond to dataset type.
│ Appending '.vtu' to filename.
└ @ WriteVTK ~/.julia/packages/WriteVTK/H5smu/src/WriteVTK.jl:162
┌ Warning: detected extension '.LidDriven_1_3' does not correspond to dataset type.
│ Appending '.vtu' to filename.
└ @ WriteVTK ~/.julia/packages/WriteVTK/H5smu/src/WriteVTK.jl:162
┌ Warning: detected extension '.LidDriven_1_4' does not correspond to dataset type.
│ Appending '.vtu' to filename.
└ @ WriteVTK ~/.julia/packages/WriteVTK/H5smu/src/WriteVTK.jl:162
┌ Warning: detected extension '.LidDriven_2_1' does not correspond to dataset type.
│ Appending '.vtu' to filename.
└ @ WriteVTK ~/.julia/packages/WriteVTK/H5smu/src/WriteVTK.jl:162
┌ Warning: detected extension '.LidDriven_2_2' does not correspond to dataset type.
│ Appending '.vtu' to filename.
└ @ WriteVTK ~/.julia/packages/WriteVTK/H5smu/src/WriteVTK.jl:162
┌ Warning: detected extension '.LidDriven_2_3' does not correspond to dataset type.
│ Appending '.vtu' to filename.
└ @ WriteVTK ~/.julia/packages/WriteVTK/H5smu/src/WriteVTK.jl:162
┌ Warning: detected extension '.LidDriven_2_4' does not correspond to dataset type.
│ Appending '.vtu' to filename.
└ @ WriteVTK ~/.julia/packages/WriteVTK/H5smu/src/WriteVTK.jl:162
[ Info: model read completed
[ Info: boundary conditions created
[ Info: FE Spaces Created
New tag centre added to the model
[ Info: model read completed
[ Info: boundary conditions created
[ Info: FE Spaces Created
[ Info: model read completed
[…]
