### 2025-01-27 #7564 @marinlauber
@carlodev

The documentation and text updates are very nice and clear up how to run the cases for me. Thanks!

Some of the comments remain; see https://github.com/openjournals/joss-reviews/issues/7564#issuecomment-2578100130

Additionally, I have a few more:

1. The file `examples/3D/Airfoil3D.jl` is empty
2. WaterLily.jl is not a Lattice-Boltzmann method; it's a CPU/GPU finite-volume method that uses an immersed-boundary method ;)

There are still two typos in the paper:

_l.54_    It has a suite of tools for post-processing the results like performing time and spanwise
_l.55_    ... averaging, **obtaining.**
_l.56_    ..... The user can load **their** own airfoil meshes also **usign**
