+++ test/sequential/runtests.jl
@@ -6,4 +6,3 @@ using Test
 
 @testset "Sequential" begin include( joinpath("SequentialTests.jl")) end
 
-@testset "Utils" begin  include(joinpath("..","UtilsTests","runtests.jl")) end
