+++ test/UtilsTests/CreateVtuTests.jl
@@ -5,7 +5,7 @@ using DataFrames
 using SegregatedVMSSolver.CreateVtu
 
 
-fname = create_vtu_file("Results_vtu")
+fname = create_vtu_file(joinpath(@__DIR__,"Results_vtu"))
 @test typeof(fname) == String
 
 end #end module
