+++ test/runtests.jl
@@ -5,7 +5,6 @@ using Test
 
 @testset "Sequential" begin  include("sequential/runtests.jl") end
 
-@testset "Utils" begin  include("UtilsTests/runtests.jl") end
 
 
 
