### 2025-01-29 #7564 @weymouth
I think the text is a bit unclear. The readme states the contributions as:
- SUPG finite element solver
- VMS turbulence model
- MPI execution
But the paper adds a lot of stuff about "empowering users with different CFD needs".

IMO, the bullet points would be sufficient - I just want to see evidence that these have been implemented correctly. Fig 1 is qualitative other than the friction plot, which doesn't have a clear "right answer". Maybe add the 3D TGV dissipation since we know the correct DNS? Also the paper is missing any documentation of the MPI execution such as a speed-up plot as the number of cores is changed.

If the user interface stuff really is a contribution, you'll need to explain it better and provide evidence. I don't know what is "out of the box" with gridapp - so I can't immediately appreciate what is different here. I also don't think that required - but in that case this claim should be toned down in the text.

The text also mentions the "gap" being the lack of LES solvers in Julia - but WaterLily is a documented LES code, and I'm pretty sure Oceananigans is as well. Better to stick with the bullet points.
