### 2025-02-05 db75aac
add cross_terms in VMS
