+++ src/Commons/Equations/StabilizedEquations.jl
@@ -6,44 +6,91 @@ Wrapper for the convective term
 """
 cconv(u, ∇u) = (∇u') ⋅ u
 
+
+function Rm_adv_update(un,un_1,pn,ν,dt, U, method)
+  _, vms_cross_activation = VMS_activation(method)
+  if vms_cross_activation 
+    Rm = (un-un_1) ./ dt +  cconv(un, ∇(un)) + ∇(pn) - ν*Δ(un)
+    return interpolate(Rm,U)
+  else
+    return nothing
+  end
+end
+
+
+macro create_equation(name, vms_flag, vms_cross_flag, base_term, vms_term, vms_cross_term)
+  func_name = esc(name)
+
+  quote
+      function $(func_name)(a, b)
+          if $(esc(vms_flag)) && $(esc(vms_cross_flag))
+              $(esc(base_term))(a, b) + $(esc(vms_term))(a, b) + $(esc(vms_cross_term))(a, b)
+          elseif $(esc(vms_flag))
+              $(esc(base_term))(a, b) + $(esc(vms_term))(a, b)
+          else
+              $(esc(base_term))(a, b)
+          end
+      end
+  end
+end
+
+
+
 function segregated_equations(u_adv,params::Dict{Symbol,Any},simcase::SimulationCase)
   @sunpack skew, ν,dt, θ,D = simcase
   
-  sprob = simcase.sprob
-  @unpack dΩ = params
-  @unpack skew = sprob
-    
+    sprob = simcase.sprob
+    @unpack dΩ = params
+    @unpack skew = sprob
+      
     skewcoeff = skew * 0.5 # ==0 if skew == false
-    vms_activation = is_VMS(sprob.method)
+    vms_activation, vms_cross_activation = VMS_activation(sprob.method)
     
-    visc_term =   0.0 #- ν* compute_Δ(u_adv,D) # compute the laplacian of the previous time step
     stab_coeff = compute_stab_coeff(simcase,params)
     Tm = momentum_stabilization(u_adv, stab_coeff, simcase)
     Tc = continuity_stabilization(u_adv, stab_coeff, simcase)
-  
-    ### VMS extra equations
-    Auu_vms1(u, v) =  ∫(vms_activation .* u_adv ⋅ (∇(v))'*Tm⊙ (cconv ∘ (u_adv, ∇(u)) ) )dΩ + 
-     ∫((vms_activation *skewcoeff) .* u_adv ⋅ (∇(v))'*Tm⊙ (u_adv ⋅ (∇ ⋅ u)) )dΩ
+    
 
-    Tuu_vms1(u, v) =  ∫(vms_activation .* u_adv ⋅ (∇(v))'*Tm⊙u )dΩ # VMS1
+      if vms_cross_activation
+        @unpack Rm_adv = params
+      end
 
[…]
