### 2025-02-11 3be2f50
agglomerate equations integrals
