+++ src/Commons/ExportUtility.jl
@@ -365,7 +365,7 @@ function print_on_request(log_dir::String)
 end
 
 
-function writesolution(params::Dict{Symbol,Any}, simcase::SimulationCase, ntime::Int64, tn::Float64, fields::Tuple)
+function writesolution(params::Dict{Symbol,Any}, simcase::SimulationCase, ntime::Int64, tn::Float64, fields::Tuple, avg_fields::Tuple)
     benchmark = simcase.simulationp.exportp.benchmark
     log_dir = simcase.simulationp.exportp.log_dir
     nsub = simcase.sprob.method.order
@@ -375,99 +375,186 @@ function writesolution(params::Dict{Symbol,Any}, simcase::SimulationCase, ntime:
             save_sim_dir = simcase.simulationp.exportp.save_sim_dir
             case = typeof(simcase)
             save_path = joinpath(save_sim_dir, "$(case)_$(tn)_.vtu")
-            @unpack Ω = params
-            writesolution(simcase, Ω, nsub, save_path, tn, fields)
+        
+            writesolution(simcase, params, nsub, save_path, tn, fields,avg_fields)
         end
-        compute_error(params, simcase, tn, fields)
+        compute_enstrophy_ek(simcase, params,fields,tn)
+
     end
 
 end
 
 
-function writesolution(simcase::TaylorGreen{Periodic}, Ω, nsub::Int64, save_path, tn::Float64, fields::Tuple)
-    u_analytic = simcase.bc_type.a_solution[:velocity]
-    p_analytic = simcase.bc_type.a_solution[:pressure]
+function writesolution(simcase, params::Dict{Symbol,Any}, nsub::Int64, save_path::String, tn::Float64, fields::Tuple, avg_fields::Tuple)
     uh, ph = fields
+    uh_avg, ph_avg =avg_fields
+    fields_names = simcase.simulationp.exportp.vtu_export
+    @unpack Ω = params
 
-    writevtk(Ω, save_path, nsubcells=nsub, cellfields=["uh" => uh, "uh_analytic" => u_analytic(tn), "ph" => ph, "ph_analytic" => p_analytic(tn)])
+    ALLOWED_EXPORTS = ["uh", "ph", "uh_avg", "ph_avg"]
+    selected_exports = intersect(fields_names,ALLOWED_EXPORTS)
+    cell_fields = Dict{String,Any}()
+    if "uh" in selected_exports
+        merge!(cell_fields,Dict("uh"=>uh))
+    end
+    if "ph" in selected_exports
+        merge!(cell_fields,Dict("ph"=>ph))
+    end
+    if "uh_avg" in selected_exports
+        merge!(cell_fields,Dict("uh_avg"=>uh_avg))
+    end
+    if "ph_avg" in selected_exports
+        merge!(cell_fields,Dict("ph_avg"=>ph_avg))
+    end
+   
+    writesolution(simcase, tn, fields_names,cell_fields)
+    compute_error(simcase, params,   tn, fields)
+
+    writevtk(Ω, save_path, nsubcells=nsub, cellfields=cell_fields)
 end
 
-
-function writesolution(simcase::TaylorGreen{Natural}, Ω, nsub::Int64, save_path, tn::Float64, fields::Tuple)
-    uh, ph = fields
-    writevtk(Ω, save_path, nsubcells=nsub, cellfields=["uh" => uh, "ph" => ph])
+function writesolution(simcase, tn::Float64, fields_names, cell_fields::Dict{String,Any})
 end
 
-
-function writesolution(simcase::VelocityBoundaryCase, Ω, nsub::Int64, save_path, tn, fields::Tuple)
-    uh_tn, ph_tn, uh_tn_updt, uh_avg, ph_avg = fields
[…]
