+++ docs/make.jl
@@ -18,7 +18,8 @@ makedocs(;plugins=[bib],
         "MPI run" => "mpi.md",
         "Examples" =>[
         "Intro"=>"Cases/examples.md",
-        "Taylor Green" => "Cases/taylorgreen.md",
+        "Taylor Green 2D" => "Cases/taylorgreen2D.md",
+        "Taylor Green 3D" => "Cases/taylorgreen3D.md",
         "Lid Driven Cavity Flow" => "Cases/liddriven.md",        
         "Cylinder" => "Cases/cylinder.md",
         "Airfoil" => "Cases/airfoil.md",
