+++ docs/src/Cases/airfoil.md
@@ -2,7 +2,7 @@
 
 ![LSB-s7003](../assets/LSB-sd7003.png)
 
-It is one of the most complex and intersting case. The user has to create a proper mesh in [`gmsh`](https://gmsh.info/) setting the following physical boundaries:
+It is one of the most complex and interesting case. The user has to create a proper mesh in [`gmsh`](https://gmsh.info/) setting the following physical boundaries:
 - `inlet` for the inlet
 - `outlet` for the outlet
 - `airfoil` for the airfoil walls
@@ -13,64 +13,64 @@ It is one of the most complex and intersting case. The user has to create a prop
     In the folder `models/geo` are available some `.geo` that can be open in [`gmsh`](https://gmsh.info/).
     Different parameters can be modified: angle of attack, domain dimension, mesh divisions.
 
-The velocity at the inlet is incresed from `0.0` arriving to the target value `u_in` at `:t_endramp`. This increase the numeric stability. If `:t_endramp` = `:t0` the velocity at the inlet will be immediately `:u_in`. For numeric stability is better to keep `u_in = [1.0,0.0,0.0]`, then fix the Reynolds and so the viscosity will be automatically computed as: `ν = 1/Reynolds`
+The velocity at the inlet is increased from `0.0` arriving to the target value `u_in` at `t_endramp`. This increase the numeric stability. If `t_endramp` = `t0` the velocity at the inlet will be immediately `u_in`. For numeric stability is better to keep `u_in = [1.0,0.0,0.0]`, then fix the Reynolds and so the viscosity will be automatically computed as: `ν = 1/Reynolds`
 
 The pressure is set `0.0` at the `outlet` section. The velocity on the `limits` is set equal to the one at `inlet`.
 
-## Suggested workflow
-3D LES are heavy and it is possible to experience divergence issues. It is suggested to use one of the two initilization tecniques: `Velocity ramping` or `Boundary layer initialization`. For visualize the results, please refer to 
+3D LES are heavy and it is possible to experience divergence issues. It is suggested to use one of the two initialization techniques: `Velocity ramping` or `Boundary layer initialization`. For visualize the results, please refer to the section. 
 
-### Velocity ramping
-By setting `t_endramp > t0` automatically the code will create an inlet velocity which will increase linearly in time up to the `u_in` target value.
 
+## NACA0012
+A common validation test case is to analyze the NACA0012 airfoil and compare the pressure distribution with experimental results. This is a simple 2D simulation at high Reynolds number ``3e6``. For run this case it is suggested to have `hypre` installed to use a more powerful preconditioner for the resolution. It has been tested on 16 CORES and requires approximately 35 hours.
+
+
+### Simulation File
 ```julia
 using PartitionedArrays
-using MPI
 using SegregatedVMSSolver
 using SegregatedVMSSolver.ParametersDef
 using SegregatedVMSSolver.SolverOptions
 ```
-Calling the dependencies
 
-
+Set the physical parameters of the simulation
 ```julia
 t0 =0.0
-dt = 1e-3
-tF = 5e-3
+dt = 0.00025
+tF = 15.0
 
-Re = 10
+Re = 3_000_000
 D = 2
-rank_partition = (2,2)
[…]
