+++ .github/workflows/CI.yml
@@ -3,6 +3,7 @@ on:
   push:
     branches:
       - master
+      - JossPaper
     tags: ['*']
 jobs:
   REPLTestsCommons:
