+++ paper/paper.bib
@@ -79,3 +79,7 @@
    title = {Development of Stabilized Finite Element Method for Numerical Simulation of Turbulent Incompressible Single and Eulerian-Eulerian Two-Phase Flows},
    year = {2016},
 }
+
+
+ @article{botella_peyret:1998, 
+ title={Benchmark spectral results on the lid-driven cavity flow}, volume={27}, DOI={10.1016/s0045-7930(98)00002-4}, number={4}, journal={Computers and Fluids}, author={Botella, O. and Peyret, R.}, year={1998}, pages={421–433}} 
