+++ docs/make.jl
@@ -28,7 +28,8 @@ makedocs(;plugins=[bib],
         "Theory"=>[
             "Incompressible Navier Stokes"=>"Theory/NS_inc.md",
             "FEM"=>"Theory/FEMTheory.md",
-            "SUPG & VMS Stabilization"=>"Theory/SUPG_VMS_stab.md"],
+            "SUPG & VMS Stabilization"=>"Theory/SUPG_VMS_stab.md",
+            "Error and Parallelization"=>"Theory/Errors.md"],
         "Tools"=>[       
         "Visualization" => "Tools/visualization.md",
         "Post Processing" => "Tools/post_proc.md",
