+++ docs/src/Cases/taylorgreen3D.md
@@ -79,7 +79,7 @@ at each time-step, a new line on a `.csv` file will be written with: timestep-Ki
 The Linearized and Segregated VariationalMultiScale method has been validated with test case. The time-step is constant ``\Delta t =0.01s``, the matrices are re-computed every 5 time-steps. The Reynolds number is ``1600``, it is used to compute the viscosity: ``\nu = U_0 k / Re``. Order of interpolation of velocity and pressure is the same. The benchmark is the DNS result. The flow patterns generated by the Taylor-Green vortex are quite complex. The resolution of the most complex case tested: 64x64x64 Q2Q2, takes 40 hours on a 64CORE machine.
 
 We note that the kinetic energy decay does not match that of the reference for Q1 elements, there is significant numerical dissipation since the enstrophy does not match the kinetic energy decay.
-![TGV](../assets/TGV_64_Q1.png)
+![TGV](../assets/TGV_64_Q1.png )
 
 Increasing the order from Q1 to Q2 yield the following results which are better.
 ![TGV](../assets/TGV_32_Q2.png)
