+++ .github/workflows/CI.yml
@@ -3,7 +3,6 @@ on:
   push:
     branches:
       - master
-      - JossPaper
     tags: ['*']
 jobs:
   REPLTestsCommons:
