+++ .github/workflows/CI.yml
@@ -41,7 +41,7 @@ jobs:
           path: ~/.julia/artifacts
           key: ubuntu-test-cache-${{ hashFiles('**/Project.toml') }}
       - uses: julia-actions/julia-buildpkg@v1
-      - run: julia --project=. -e "include(joinpath('test', 'sequential', 'runtests.jl'))"
+      - run: julia --project=. -e "include(joinpath(\"test\", \"sequential\", \"runtests.jl\"))"
       - uses: julia-actions/julia-processcoverage@v1
       - uses: actions/upload-artifact@v3
         with:
@@ -62,7 +62,7 @@ jobs:
           path: ~/.julia/artifacts
           key: ubuntu-test-cache-${{ hashFiles('**/Project.toml') }}
       - uses: julia-actions/julia-buildpkg@v1
-      - run: julia --project=. -e "include(joinpath('test', 'UtilsTests', 'runtests.jl'))"
+      - run: julia --project=. -e "include(joinpath(\"test\", \"UtilsTests\", \"runtests.jl\"))"
       - uses: julia-actions/julia-processcoverage@v1
       - uses: actions/upload-artifact@v3
         with:
@@ -83,7 +83,7 @@ jobs:
           path: ~/.julia/artifacts
           key: ubuntu-mpi-test-cache-${{ hashFiles('**/Project.toml') }}
       - uses: julia-actions/julia-buildpkg@v1
-      - run: julia --project=. -e "include(joinpath('test', 'mpi', 'runtests.jl'))"
+      - run: julia --project=. -e "include(joinpath(\"test\", \"mpi\", \"runtests.jl\"))"
 
   mergeCoverage:
     name: Merge Coverage & Upload
