+++ .github/workflows/CI.yml
@@ -91,7 +91,7 @@ jobs:
     needs: [REPLTestsCommons, REPLTestsCases, UtilsTests]
     steps:
       - uses: actions/checkout@v3
-      - uses: actions/download-artifact@v3
+      - uses: actions/download-artifact@v4
         with:
           path: coverage_reports
       - run: |
