+++ .github/workflows/CI.yml
@@ -95,9 +95,7 @@ jobs:
         with:
           path: coverage_reports
       - run: julia --project=. -e 'import Pkg; Pkg.add("Coverage")'  # Install Coverage.jl
-      - run: julia --project=. -e 'using Coverage; files = filter(f -> endswith(f, ".info"), readdir("coverage", join=true));
-            merged_cov = merge_coverage_counts([LCOV.read(f) for f in files]);  # Correct function
-            LCOV.write("coverage/lcov.info", merged_cov);'
+      - run: julia --project=. -e 'using Coverage; files = filter(f -> endswith(f, ".info"), readdir("coverage", join=true)); merged_cov = merge_coverage_counts([LCOV.read(f) for f in files]);   LCOV.write("coverage/lcov.info", merged_cov);'
       - uses: codecov/codecov-action@v2
         with:
           token: ${{ secrets.CODECOV_TOKEN }}
