### 2025-02-27 #7564 @carlodev
The documentation has now been updated to reflect the latest changes. In addition, we are running multiple tests in parallel, significantly reducing the overall testing time, and the reported coverage is the combined result from all parallel jobs. These improvements build on the implementation reported last week. I am available for any clarifications, new implementation suggestions, or to address any issues you may encounter.
