### 2025-03-10 09bdb33
Adding Fully Coupled Non-Linear Case
