### 2025-03-16 #7564 @weymouth
I'm happy with the proof. nice work @carlodev
