### 2025-03-18 #7564 @nathanquinlan
Hi @carlodev, sorry for the delay. I am unable to view results from simulations when running examples out of the box (Taylor-Green etc). The /Results directory is created but is empty. Maybe I'm missing something simple. Also, the visualization doc page appears to be a duplicated of part of the postprocessing page.
