### 2025-03-21 #7564 @nathanquinlan
> [@nathanquinlan](https://github.com/nathanquinlan) thanks for your feedback. I have fixed the duplicate documentation (in the `dev` version), thanks for pointing it out.
>
> For example, in `TaylorGreen2D.jl` you can uncomment lines 52-54, so a file in `Log/PrintSim.tex` is created and then `.vtu` and `.pvtu` files are saved on disks at each time-step. Otherwise you can create the folder and file manually. I came up with this solution because for 3D simulations the files can take up a lot of space and "communicate" with the script while it is running also in MPI. Actually, `.pvtu` files are saved in the folder `Results_vtu/`. In `Results/` folder are exported `.csv` files storing files from specified boundaries. From the example `Airfoil2D.jl`
>
> `exportp = ExportParameters(printinitial=true,printmodel=true,name_tags=["airfoil"], fieldexport=[["uh","ph","friction"]]) ` At each time-step, for the `airfoil` boundary, the velocity, pressure and friction field are exported, in order to reduce the allocation on data on the disks.
>
> If you have other problems or suggestions I am glad to fix them or implement new features

Thanks @carlodev. And I think you made some changes to the examples in this area that I missed. I am ready to sign off my review. I'll update my checklist now.
