### 2025-03-24 carlodev/SegregatedVMSSolver.jl#24 @marinlauber
@carlodev the test now pass on the main branch, thanks. I'll close this one as well.
