### 2025-03-24 #7564 @Kevin-Mattheus-Moerman
Thanks @marinlauber the performance box is not ticked. Are you able to, or is there an issue there still? (Or do you feel you cannot evaluate performance?)
