### 2025-04-04 #7564 @marinlauber
Sorry, I didn't see that the performance box was not ticked. All good now.
Thanks.
