### 2025-05-05 #7564 @carlodev
Sure, here the final required steps:
zenodo archive 10.5281/zenodo.15340506
new version published and tagged on the repo: v3.0.1

Let me know if there is any other step or action required from my side, thanks in advance
