### 2025-05-09 #7564 @Kevin-Mattheus-Moerman
@carlodev the archive lists the version as `Version v1`, this should become `Version v3.0.1`. Can you update the Zenodo archive listed version to match please?
