### 2025-05-19 19140f0
Fix LaTeX math expression for JATS compatibility (replace \dfrac with \frac)
