### 2025-05-19 #7564 @carlodev
I believe I’ve just fixed the issue flagged by the editorial bot — it was due to a \dfrac in an inline math expression, which I’ve now replaced with \frac.
