### 2025-05-19 #7564 @xuanxu
yes, using `\frac` should fix the error
