### 2025-05-23 a5f04b4
Updates JOSS paper
