+++ paper/paper.md
@@ -29,12 +29,13 @@ However, for more advanced users who may desire greater control over the simulat
 
 
 # Statement of need
-`SegregatedVMSSolver.jl` is a comprehensive Julia package designed for conducting high-fidelity simulations of complex flow phenomena within the incompressible regime, leveraging the VMS and SUPG method. VMS has been originally introduced in [@Hughes:2000]. The linearization adopted has been proposed in the SUPG method in [@Banyai:2016]. 
-The package relies on `Gridap.jl`[@Verdugo:2022],[@Badia:2020] package to implement the mathematical model of FEM. Complementing this core functionality, `GridapDistributed.jl`[@BadiaD:2022] and [`PartitionedArrays.jl`](https://github.com/fverdugo/PartitionedArrays.jl) allow the use of multi-core CPUs desktop computers to HPC clusters. The [`GridapPETSc.jl`](https://github.com/gridap/GridapPETSc.jl) package is used to solve the final linear system. 
+`SegregatedVMSSolver.jl` is a comprehensive Julia package designed for conducting high-fidelity simulations of complex flow phenomena within the incompressible regime, leveraging the VMS and SUPG method. VMS has been originally introduced by @Hughes:2000. The linearization adopted has been proposed in the SUPG method by @Banyai:2016. 
+The package relies on the `Gridap.jl` package [@Verdugo:2022; @Badia:2020] to implement the mathematical model of FEM. Complementing this core functionality, `GridapDistributed.jl` [@BadiaD:2022] and [`PartitionedArrays.jl`](https://github.com/fverdugo/PartitionedArrays.jl) allow the use of multi-core CPUs desktop computers to HPC clusters. The [`GridapPETSc.jl`](https://github.com/gridap/GridapPETSc.jl) package is used to solve the final linear system. 
 It solves a Linearized and Segregated version of VMS (LS-VMS) and SUPG. It is based on an iterative predictor-corrector scheme and it avoids the resolution of a non-linear problem. It uses the $\theta$ method to solve the time-marching problem. 
 
 This package focuses on solving fluid dynamics problems, with a particular emphasis on the study of 2D and 3D airfoil aerodynamics at low Reynolds numbers (tested up to 500,000). The primary objective is to provide researchers and engineers with a versatile tool for analyzing aerodynamic features in this regime. 
 A suite of benchmark cases is implemented to ensure reliability and applicability across a range of scenarios. These include:
+
 - **Taylor-Green vortices**: Simulated in both 2D and 3D.
 - **Lid-driven cavity**: Solvable in 2D; the 3D version is not currently supported.
 - **Vortex shedding over a cylinder**: Available in both 2D and 3D.
@@ -52,7 +53,7 @@ While these packages vary in scope and functionality, they collectively highligh
 
 
 # Package Features
-- It supports 2-dimensional and 3-dimensional geometries
+- It supports two-dimensional and three-dimensional geometries
 - It solved a time-dependent problem
 - It can run in parallel using MPI
 - Velocity ramping
@@ -65,18 +66,23 @@ While these packages vary in scope and functionality, they collectively highligh
 # Results
 
 ## Taylor Green 2D
-2D Taylor-Green vortex case is used as benchmark case to assess the error analysis of the code. The domain is a square of size: [-0.5,0.5]x[-0.5,0.5], at Reynolds number 1600. The CFL is constant for each simulation, CFL=0.32. The time-step is computed as $dt = \frac{CFL}{ order \cdot N}$.
+2D Taylor-Green vortex case is used as benchmark case to assess the error analysis of the code. The domain is a square of size: $[-0.5,0.5] \times [-0.5,0.5]$, at Reynolds number 1600. 
+The CFL is constant for each simulation at 0.32. 
[…]
