### 2025-05-23 0205e69
Updates JOSS references
