+++ paper/paper.bib
@@ -7,7 +7,7 @@
   number = {52},
   pages = {2520},
   author = {Santiago Badia and Francesc Verdugo},
-  title = {Gridap: An extensible Finite Element toolbox in Julia},
+  title = {Gridap: An extensible Finite Element toolbox in {J}ulia},
   journal = {Journal of Open Source Software}
 }
 
@@ -20,7 +20,7 @@
   volume = {276},
   pages = {108341},
   author = {Francesc Verdugo and Santiago Badia},
-  title = {The software design of Gridap: A Finite Element package based on the Julia {JIT} compiler},
+  title = {The software design of {Gridap}: A Finite Element package based on the {J}ulia {JIT} compiler},
   journal = {Computer Physics Communications}
 }
 
@@ -33,7 +33,7 @@
   number = {74},
   pages = {4157},
   author = {Santiago Badia and Alberto F. Martín and Francesc Verdugo},
-  title = {GridapDistributed: a massively parallel finite element toolbox in Julia},
+  title = {GridapDistributed: a massively parallel finite element toolbox in {J}ulia},
   journal = {Journal of Open Source Software}
 }
 
@@ -72,14 +72,26 @@
   doi = {10.1007/s007910050051},
 }
 
-@article{Banyai:2016,
+@phdthesis{Banyai:2016,
    author = {Tamas Banyai},
    isbn = {9782875161093},
-   institution = {Université Libre de Bruxelles},
-   title = {Development of Stabilized Finite Element Method for Numerical Simulation of Turbulent Incompressible Single and Eulerian-Eulerian Two-Phase Flows},
+   school = {Université Libre de Bruxelles, Ecole polytechnique de Bruxelles},
+   address = {Bruxelles},
+   title = {Development of Stabilized Finite Element Method for Numerical Simulation of Turbulent Incompressible Single and {E}ulerian-{E}ulerian Two-Phase Flows},
    year = {2016},
+   type = {PhD thesis},
+   month = aug,
+   url = {http://hdl.handle.net/2013/ULB-DIPOT:oai:dipot.ulb.ac.be:2013/235110}
 }
 
 
- @article{botella_peyret:1998, 
- title={Benchmark spectral results on the lid-driven cavity flow}, volume={27}, DOI={10.1016/s0045-7930(98)00002-4}, number={4}, journal={Computers and Fluids}, author={Botella, O. and Peyret, R.}, year={1998}, pages={421–433}} 
+@article{botella_peyret:1998, 
+  title={Benchmark spectral results on the lid-driven cavity flow}, 
+  volume={27}, 
+  DOI={10.1016/s0045-7930(98)00002-4}, 
+  number={4}, 
+  journal={Computers and Fluids}, 
+  author={Botella, O. and Peyret, R.}, 
+  year={1998}, 
+  pages={421–433}
+}
