### 2025-05-23 #7564 @kyleniemeyer
Hi @carlodev, I've made a few edits to the paper in https://github.com/carlodev/SegregatedVMSSolver.jl/pull/33, could you review and merge those?

Also, I noticed that you used hyperlinks for a number of software packages in the paper:
- PartitionedArrays.jl
- GridapPETSc.jl
- AirfoilGmsh.jl
- Trixi.jl
- WaterLily.jl
- Oceananigans.jl

Please replace those links with direct citations to the packages; many/most have DOIs you can cite along with including the GitHub URLs.
