+++ paper/paper.bib
@@ -95,3 +95,44 @@
   year={1998}, 
   pages={421–433}
 }
+
+
+
+@misc{Trixi:2025,
+  title={{T}rixi.jl: {A}daptive high-order numerical simulations
+         of hyperbolic {PDE}s in {J}ulia},
+  author={Schlottke-Lakemper, Michael and Gassner, Gregor J and
+          Ranocha, Hendrik and Winters, Andrew R and Chan, Jesse
+          and Rueda-Ramírez, Andrés},
+  year={2025},
+  howpublished={\url{https://github.com/trixi-framework/Trixi.jl}},
+  doi={10.5281/zenodo.3996439}
+}
+
+
+
+@misc{WaterLily:2024,
+      title={WaterLily.jl: A differentiable and backend-agnostic Julia solver to simulate incompressible viscous flow and dynamic bodies}, 
+      author={Gabriel D. Weymouth and Bernat Font},
+      year={2024},
+      eprint={2407.16032},
+      archivePrefix={arXiv},
+      primaryClass={physics.flu-dyn},
+      url={https://arxiv.org/abs/2407.16032}, 
+}
+
+
+@article{Oceananigans:2020,
+ doi = {10.21105/joss.02018}, url = {https://doi.org/10.21105/joss.02018}, year = {2020}, publisher = {The Open Journal}, volume = {5}, number = {53}, pages = {2018}, author = {Ali Ramadhan and Gregory LeClaire Wagner and Chris Hill and Jean-Michel Campin and Valentin Churavy and Tim Besard and Andre Souza and Alan Edelman and Raffaele Ferrari and John Marshall}, title = {Oceananigans.jl: Fast and friendly geophysical fluid dynamics on GPUs}, journal = {Journal of Open Source Software} }
+
+
+
+ @misc{Verdugo:2025,
+  author       = {Francesc Verdugo and Alberto F. Martin and HRvanElderen and GeliezaK and Jordi Manyer Fuertes and Fredrik Ekre and Joroks and Dennis Ogiermann and Yung-sheng Tu and Hendrik Ranocha and Oriol Colomés and Mosè Giordano and Oscar Smith and Robin Richardson},
+  title        = {PartitionedArrays.jl: PartitionedArrays-v0.5.10},
+  year         = 2025,
+  publisher    = {Zenodo},
+  version      = {v0.5.10},
+  doi          = {10.5281/zenodo.15266760},
+  url          = {https://github.com/PartitionedArrays/PartitionedArrays.jl}
+}
