### 2025-05-23 #7564 @carlodev
Thank you very much for your helpful edits—they are greatly appreciated. I've added formal citations for the packages you mentioned, using DOIs where available. For GridapPETSc.jl and AirfoilGmsh.jl, I could not find any associated publications or Zenodo records to cite, so I have retained the GitHub URLs for those.
