### 2025-05-23 #7564 @kyleniemeyer
@carlodev thank you - but even for the software packages without DOIs, you can still cite the packages (with author names, etc.) like regular references. DOIs are just helpful if available. (This is important because we want to treat software at the same level as other scholarly works, even if they do not have an associated paper or formal publication.)
