### 2025-05-23 9edc555
add papers references
