+++ paper/paper.bib
@@ -136,3 +136,22 @@
   doi          = {10.5281/zenodo.15266760},
   url          = {https://github.com/PartitionedArrays/PartitionedArrays.jl}
 }
+
+
+
+@misc{@BrunelliAirfoilGmsh:2023,
+  author       = {Carlo Brunelli},
+  title        = {AirfoilGmsh.jl: A Julia package for generating airfoil meshes using Gmsh},
+  year         = {2023},
+  howpublished = {\url{https://github.com/carlodev/AirfoilGmsh.jl}},
+  note         = {Accessed: 2025-05-23}
+}
+
+
+@misc{GridapPETSc:2025,
+  author       = {Francesc Verdugo and Santiago Badia and Alberto F. Martin and Jordi Manyer Fuertes and Eric Neiva and the Gridap Team},
+  title        = {GridapPETSc.jl: A PETSc backend for Gridap.jl},
+  year         = {2025},
+  howpublished = {\url{https://github.com/gridap/GridapPETSc.jl}},
+  note         = {Accessed: 2025-05-23}
+}
