+++ paper/paper.bib
@@ -139,12 +139,11 @@
 
 
 
-@misc{@BrunelliAirfoilGmsh:2023,
+@misc{@BrunelliGmsh:2025,
   author       = {Carlo Brunelli},
   title        = {AirfoilGmsh.jl: A Julia package for generating airfoil meshes using Gmsh},
-  year         = {2023},
+  year         = {2025},
   howpublished = {\url{https://github.com/carlodev/AirfoilGmsh.jl}},
-  note         = {Accessed: 2025-05-23}
 }
 
 
@@ -153,5 +152,4 @@
   title        = {GridapPETSc.jl: A PETSc backend for Gridap.jl},
   year         = {2025},
   howpublished = {\url{https://github.com/gridap/GridapPETSc.jl}},
-  note         = {Accessed: 2025-05-23}
 }
