+++ paper/paper.bib
@@ -137,8 +137,6 @@
   url          = {https://github.com/PartitionedArrays/PartitionedArrays.jl}
 }
 
-
-
 @misc{@BrunelliGmsh:2025,
   author       = {Carlo Brunelli},
   title        = {AirfoilGmsh.jl: A Julia package for generating airfoil meshes using Gmsh},
