+++ paper/paper.bib
@@ -137,17 +137,16 @@
   url          = {https://github.com/PartitionedArrays/PartitionedArrays.jl}
 }
 
-@misc{@BrunelliGmsh:2025,
-  author       = {Carlo Brunelli},
-  title        = {AirfoilGmsh.jl: A Julia package for generating airfoil meshes using Gmsh},
-  year         = {2025},
-  howpublished = {\url{https://github.com/carlodev/AirfoilGmsh.jl}},
-}
-
-
 @misc{GridapPETSc:2025,
   author       = {Francesc Verdugo and Santiago Badia and Alberto F. Martin and Jordi Manyer Fuertes and Eric Neiva and the Gridap Team},
   title        = {GridapPETSc.jl: A PETSc backend for Gridap.jl},
   year         = {2025},
   howpublished = {\url{https://github.com/gridap/GridapPETSc.jl}},
+}
+
+@misc{@BrunelliGmsh:2025,
+  author       = {Carlo Brunelli},
+  title        = {AirfoilGmsh.jl: A Julia package for generating airfoil meshes using Gmsh},
+  year         = {2025},
+  howpublished = {\url{https://github.com/carlodev/AirfoilGmsh.jl}},
 }
