+++ paper/paper.bib
@@ -145,8 +145,4 @@
 }
 
 @misc{@BrunelliGmsh:2025,
-  author       = {Carlo Brunelli},
-  title        = {AirfoilGmsh.jl: A Julia package for generating airfoil meshes using Gmsh},
-  year         = {2025},
-  howpublished = {\url{https://github.com/carlodev/AirfoilGmsh.jl}},
-}
+  author = {Carlo Brunelli}, title={AirfoilGmsh.jl: A Julia packgage for generating airfoil meshes using Gmsh}, year={2025}, howpublished={\url{https://github.com/carlodev/AirfoilGmsh.jl}}}
