+++ paper/paper.bib
@@ -146,4 +146,3 @@
 
 @misc{@BrunelliGmsh:2025,
   author = {Carlo Brunelli}, title={AirfoilGmsh.jl: A Julia packgage for generating airfoil meshes using Gmsh}, year={2025}, howpublished={\url{https://github.com/carlodev/AirfoilGmsh.jl}}}
-  
