### 2025-06-02 #7564 @carlodev
I have just added the citations according to your instructions
