+++ Project.toml
@@ -32,27 +32,3 @@ Trapz = "592b5752-818d-11e9-1e9a-2b8ca4a44cd1"
 UnPack = "3a884ed6-31ef-47d7-9d2a-63182c4928ed"
 
 [compat]
-CSV = "0.10.15"
-DataFrames = "1.7.0"
-DataInterpolations = "6.6.0"
-DelimitedFiles = "1.9.1"
-FFTW = "1.8.0"
-FileIO = "1.16.6"
-FillArrays ="=1.13.0"
-Gridap = "=0.17.20"
-GridapDistributed = "=0.3.2"
-GridapGmsh = "=0.7.1"
-GridapPETSc = "=0.5.3"
-MPI ="=0.20.22"
-NearestNeighbors ="0.4.21"
-Parameters = "0.12.3"
-PartitionedArrays ="=0.3.4"
-Revise ="3.7.1"
-ScatteredInterpolation = "0.3.6"
-LinearAlgebra = "1.11.0"
-SparseArrays = "1.11.0"
-SyntheticEddyMethod = "=0.4.4"
-Trapz ="2.0.3"
-Test = "1.11.0"
-UnPack = "1.0.2"
-julia = "1.11.1"
