+++ Project.toml
@@ -32,3 +32,29 @@ Trapz = "592b5752-818d-11e9-1e9a-2b8ca4a44cd1"
 UnPack = "3a884ed6-31ef-47d7-9d2a-63182c4928ed"
 
 [compat]
+CSV = "^0.10.15"
+DataFrames = "^1.7.0"
+DataInterpolations = "^8.0.1"
+DelimitedFiles = "^1.9.1"
+FFTW = "^1.9.0"
+FileIO = "^1.17.0"
+FillArrays = "^1.13.0"
+Gridap = "^0.18.12"
+GridapDistributed = "^0.4.7"
+GridapGmsh = "^0.7.2"
+GridapPETSc = "^0.5.3"
+LinearAlgebra = "^1.11.0"
+MPI = "^0.20.22"
+NearestNeighbors = "^0.4.21"
+Parameters = "^0.12.3"
+PartitionedArrays = "^0.3.4"
+Pkg = "^1.11.0"
+Random = "^1.11.0"
+Revise = "^3.8.0"
+ScatteredInterpolation = "^0.3.6"
+SparseArrays = "^1.11.0"
+Statistics = "^1.11.1"
+SyntheticEddyMethod = "^0.4.4"
+Test = "^1.11.0"
+Trapz = "^2.0.3"
+UnPack = "^1.0.2"
